% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/extract.R
\docType{methods}
\name{access}
\alias{access}
\alias{access-method}
\alias{getID}
\alias{getTotals}
\alias{getDates}
\alias{setDates}
\alias{getCoordinates}
\alias{setCoordinates}
\alias{getID,ANY-method}
\alias{getTotals,FrequencyMatrix-method}
\title{Extract or Replace Parts of an Object}
\usage{
getID(x)

getTotals(x)

getDates(x)

setDates(x)

getCoordinates(x)

setCoordinates(x)

\S4method{getID}{ANY}(x)

\S4method{getTotals}{FrequencyMatrix}(x)
}
\arguments{
\item{x}{An object from which to extract element(s).}
}
\description{
Getters and setters to extract or replace parts of an object.
}
\author{
N. Frerebeau
}
