% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{access}
\alias{access}
\alias{get}
\alias{set}
\title{Get or Set Parts of an Object}
\arguments{
\item{object}{An object from which to get or set element(s).}
}
\value{
An object of the same sort as \code{object} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\seealso{
Other mutator: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutator}
