% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{accessors}
\alias{accessors}
\alias{columns}
\alias{method}
\alias{rows}
\alias{totals}
\title{Accessors}
\usage{
columns(object)

method(object)

rows(object)

totals(object)
}
\arguments{
\item{object}{An object.}
}
\description{
Accessors
}
\author{
N. Frerebeau
}
