% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{altamira}
\alias{altamira}
\title{Altamira Engraved Bones}
\format{
A numeric matrix with 5 rows and 44 variables (designs):
}
\usage{
altamira
}
\description{
A dataset of design elements in engraved bones from Altamira (Spain).
}
\references{
Conkey, M. W. (1980). The Identification of prehistoric hunter-gatherer
 aggregation sites: The case of Altamira. \emph{Current Anthropology}, 21(5),
 609-630.
}
\seealso{
Other datasets: 
\code{\link{boves}},
\code{\link{compiegne}},
\code{\link{merzbach}},
\code{\link{mississippi}},
\code{\link{zuni}}
}
\concept{datasets}
\keyword{datasets}
