% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aves}
\alias{aves}
\title{Birds Species and Abundances}
\format{
A \code{\link{data.frame}} with 2 rows and 26 variables (bird species).
}
\source{
Magurran, A. E. (1988). \emph{Ecological Diversity and its Measurement}.
Princeton, NJ: Princeton University Press.
\doi{10.1007/978-94-015-7358-0}.
}
\usage{
aves
}
\description{
A dataset of birds species and abundances in managed and unmanaged areas
along the River Wye (UK).
}
\seealso{
Other datasets: 
\code{\link{cantabria}},
\code{\link{pueblo}},
\code{\link{woodland}}
}
\concept{datasets}
\keyword{datasets}
