% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beta-diversity.R
\docType{methods}
\name{beta-diversity}
\alias{beta-diversity}
\alias{turnover}
\alias{turnover-method}
\alias{similarity}
\alias{similarity-method}
\alias{turnover,CountMatrix-method}
\alias{turnover,FrequencyMatrix-method}
\alias{turnover,IncidenceMatrix-method}
\alias{similarity,CountMatrix-method}
\alias{similarity,FrequencyMatrix-method}
\alias{similarity,IncidenceMatrix-method}
\title{\eqn{\beta}{\beta}-diversity}
\usage{
turnover(object, ...)

similarity(object, ...)

\S4method{turnover}{CountMatrix}(object, method = c("whittaker", "cody",
  "routledge1", "routledge2", "routledge3", "wilson"), simplify = FALSE,
  ...)

\S4method{turnover}{FrequencyMatrix}(object, method = c("whittaker",
  "cody", "routledge1", "routledge2", "routledge3", "wilson"),
  simplify = FALSE, ...)

\S4method{turnover}{IncidenceMatrix}(object, method = c("whittaker",
  "cody", "routledge1", "routledge2", "routledge3", "wilson"),
  simplify = FALSE, ...)

\S4method{similarity}{CountMatrix}(object, method = c("bray", "jaccard",
  "morisita", "sorenson"), ...)

\S4method{similarity}{FrequencyMatrix}(object, method = c("jaccard",
  "sorenson"), ...)

\S4method{similarity}{IncidenceMatrix}(object, method = c("jaccard",
  "sorenson"), ...)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} matrix.}

\item{...}{Further arguments passed to other methods.}

\item{method}{A \code{\link{character}} string specifiying the method to be
used (see details). Any unambiguous substring can be given.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? The default value, \code{FALSE}, returns a list.}
}
\value{
\code{similarity} returns a \eqn{m \times \m}{m x m} symetric matrix.
 If \code{simplify} is \code{FALSE}, \code{turnover} returns a list
 (default), else returns a matrix.
}
\description{
Measures between samples diversity. \code{similarity} returns a similarity
 matrix. \code{turnover} returns the degree of turnover in taxa composition
 along a grandient or transect.
}
\details{
The following methods can be used to acertain the degree of \emph{turnover}
 in taxa composition along a gradient (\eqn{\beta}-diversity) on qualitative
 (presence/absence) data. This assumes that the order of the matrix rows
 (from 1 to \eqn{n}) follows the progression along the gradient/transect.

 \describe{
  \item{whittaker}{}
  \item{cody}{}
  \item{routledge1}{}
  \item{routledge2}{}
  \item{routledge3}{}
  \item{wilson}{}
 }

 \eqn{\beta}-diversity can also be measured by addressing \emph{similarity}
 between pairs of sites. The following methods return a value between 0
 (no similarity) and 1 (identity):
 \describe{
  \item{bray}{Sorenson quantitative index (Bray and Curtis modified version
  of the Sorenson index). }
  \item{jaccard}{Jaccard qualitative index.}
  \item{morisita}{Morisita-Horn quantitative index.}
  \item{sorenson}{Sorenson qualitative index.}
 }
}
\examples{
# Degree of turnover
# Data from Magurran 1988, p. 162
trees <- IncidenceMatrix(
  data = c(1, 1, 1, 0, 0, 0,
           1, 1, 1, 1, 1, 1,
           0, 0, 1, 0, 1, 0,
           0, 0, 0, 1, 1, 1,
           0, 0, 0, 0, 1, 1,
           0, 0, 0, 1, 0, 1),
  nrow = 6, byrow = FALSE,
  dimnames = list(c("1", "2", "3", "4", "5", "6"),
                  c("Birch", "Oak", "Rowan", "Beech", "Hazel", "Holly"))
)

## Whittaker's measure
turnover(trees, "whittaker") # 1

## Cody's measure
turnover(trees, "cody") # 3

## Routledge's measures
turnover(trees, "routledge1") # 0.29
turnover(trees, "routledge2") # 0.56
turnover(trees, "routledge3") # 1.75

## Wilson and Shmida's measure
turnover(trees, "wilson") # 1

# Similarity measures
# Data from Magurran 1988, p. 166
birds <- CountMatrix(
  data = c(1.4, 4.3, 2.9, 8.6, 4.2, 15.7, 2.0, 50, 1, 11.4, 11.4, 4.3, 13.0,
           14.3, 8.6, 7.1, 10.0, 1.4, 2.9, 5.7, 1.4, 11.4, 2.9, 4.3, 1.4, 2.9,
           0, 0, 0, 2.9, 0, 0, 0, 10, 0, 0, 5.7, 2.5, 5.7, 8.6, 5.7, 2.9, 0, 0,
           2.9, 0, 0, 5.7, 0, 2.9, 0, 2.9),
  nrow = 2, byrow = TRUE, dimnames = list(c("unmanaged", "managed"), NULL)
)

## Jaccard measure
## (presence/absence data)
similarity(birds, "jaccard") # 0.46

## Sorenson measure
## (presence/absence data)
similarity(birds, "sorenson") # 0.63

## Jaccard measure (Bray's formula)
## (count data)
similarity(birds, "bray") # 0.44

## Morisita-Horn measure
## (count data)
similarity(birds, "morisita") # 0.81
}
\seealso{
\code{\link[=richness-method]{richness}}
 \code{\link[=richness-method]{rarefaction}}
 \code{\link[=alpha-diversity]{diversity}}
 \code{\link[=alpha-diversity]{evenness}}
}
\author{
N. Frerebeau
}
