% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{BootDate-class}
\alias{BootDate-class}
\alias{[[,BootDate-method}
\alias{[,BootDate-method}
\title{Date model checking}
\usage{
\S4method{[[}{BootDate}(x, i)

\S4method{[}{BootDate}(x, i, j)
}
\arguments{
\item{x}{A \code{BootDate} object from which to extract element(s).}

\item{i, j}{Indices specifying elements to extract.
\code{i} is a \code{\link{character}} string matching to the name of a slot.
\code{j} can be \code{\link{missing}} or \code{\link{NULL}},
a \code{\link{numeric}} or \code{\link{character}} vector.
Numeric values are coerced to \code{\link{integer}} as by
\code{\link{as.integer}} (and hence truncated towards zero).
Character vectors will be matched to the names of the object.}
}
\description{
An S4 class to store the results of resampling methods for date model
 checking.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[[}: Extracts informations from a slot selected by subscript
\code{i}.

\item \code{[}: Allows to select a slot thru \code{j}.
}}

\section{Slots}{

\describe{
\item{\code{jackknife}}{A six columns \code{\link{data.frame}} giving the results of
the resamping procedure (jackknifing) for each sample (in rows) with the
following columns:
\describe{
 \item{id}{An identifier.}
 \item{estimation}{The jackknife date estimate.}
 \item{earliest}{The lower boundary of the associated prediction interval.}
 \item{latest}{The upper boundary of the associated prediction interval.}
 \item{error}{The standard error of predicted means.}
 \item{bias}{The jackknife estimate of bias.}
}}

\item{\code{bootstrap}}{A six columns \code{\link{data.frame}} giving the boostrap
distribution statistics for each replicated sample (in rows) with the
following columns:
\describe{
 \item{id}{An identifier.}
 \item{min}{Minimum value.}
 \item{Q05}{Sample quantile to 0.05 probability.}
 \item{mean}{Mean value.}
 \item{Q95}{Sample quantile to 0.95 probability.}
 \item{max}{Maximum value.}
}}
}}

\author{
N. Frerebeau
}
