% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boves}
\alias{boves}
\title{Boves ceramics}
\format{A numeric matrix with 8 rows and 13 variables:
 \describe{
   \item{I}{}
   \item{IIa}{}
   \item{IIb}{}
   \item{IIIa}{}
   \item{IIIb}{}
   \item{IIIc}{}
   \item{IVa}{}
   \item{IVb}{}
   \item{Va}{}
   \item{Vb}{}
   \item{VI}{}
   \item{VII}{}
 }}
\source{
Racinet P. (2002). Le site castral et prioral de Boves du Xe au XVIIe
 siècle. Bilan des recherches 1996-2000. \emph{Revue archéologique de
 Picardie}. Numéro spécial 20, 123 p.
}
\usage{
boves
}
\description{
A dataset containing the ceramic counts from the castle site of Boves
 (Somme, France). The data are grouped into eight periods ranging from the
 10th to the 18th century and thirteen ceramic types.
}
\keyword{datasets}
