% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CA-class}
\alias{CA-class}
\alias{.CA}
\title{Correspondence Analysis}
\description{
An S4 class to store correspondence analysis results.
}
\seealso{
Other class: 
\code{\link{BootCA-class}},
\code{\link{DateClass}},
\code{\link{DiversityIndex}},
\code{\link{PermutationOrder-class}}
}
\author{
N. Frerebeau
}
\concept{class}
