% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot_ca.R, R/run_ca.R
\docType{methods}
\name{ca}
\alias{ca}
\alias{run_ca}
\alias{run_ca-method}
\alias{predict_ca}
\alias{predict_ca-method}
\alias{plot_ca}
\alias{plot_ca-method}
\alias{plot_ca,CA-method}
\alias{plot_ca,BootCA-method}
\alias{run_ca,CountMatrix-method}
\alias{predict_ca,CA,CountMatrix-method}
\title{Correspondence Analysis}
\usage{
run_ca(object, ...)

predict_ca(object, data, ...)

plot_ca(object, ...)

\S4method{plot_ca}{CA}(object, axes = c(1, 2), map = c("rows", "columns"), ...)

\S4method{plot_ca}{BootCA}(object, axes = c(1, 2), map = c("rows", "columns"), keep = NULL, ...)

\S4method{run_ca}{CountMatrix}(object, dimensions = NULL, ...)

\S4method{predict_ca}{CA,CountMatrix}(object, data, ...)
}
\arguments{
\item{object}{A \linkS4class{CountMatrix} object.}

\item{...}{Currently not used.}

\item{data}{A \linkS4class{CountMatrix} object for which to predict
CA coordinates.}

\item{axes}{A length-two \code{\link{numeric}} vector specifying the
components to plot.}

\item{map}{A \code{\link{character}} vector specifying the graph to plot.
It must be one or both of "\code{rows}" or "\code{columns}".
Any unambiguous substring can be given.}

\item{keep}{A \code{\link{character}} vector specifying the subset to plot.
It must be \code{NULL} or one or both of "\code{rows}" or "\code{columns}".
(see \code{\link{refine_ca}}).}

\item{dimensions}{An \code{\link{integer}} value giving the number of
dimensions kept in the results. If \code{NULL} (the default), all
dimensions are kept.}
}
\description{
Performs a correspondence analysis.
}
\examples{
## Coerce dataset to a count matrix
counts <- as_count(compiegne)

## Computes correspondence analysis
ca <- run_ca(counts)
plot_ca(ca, map = c("rows", "columns"))

}
\seealso{
\link{refine_ca}, \link{seriate_correspondence}
}
\author{
N. Frerebeau
}
