% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cantabria}
\alias{cantabria}
\title{Early Magdalenian Engraved Bones}
\format{
A \code{\link{data.frame}} with 5 rows and 44 variables (designs).
}
\source{
Conkey, M. W. (1980). The Identification of prehistoric hunter-gatherer
aggregation sites: The case of Altamira. \emph{Current Anthropology}, 21(5),
609-630.

Kintigh, K. W. (1989). Sample Size, Significance, and Measures of
Diversity. In Leonard, R. D. and Jones, G. T., \emph{Quantifying Diversity
in Archaeology}. New Directions in Archaeology. Cambridge:
Cambridge University Press, p. 25-36.
}
\usage{
cantabria
}
\description{
A dataset of design elements in engraved bones from Cantabrian Spain.
}
\seealso{
Other datasets: 
\code{\link{aves}},
\code{\link{pueblo}},
\code{\link{woodland}}
}
\concept{datasets}
\keyword{datasets}
