% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\alias{checkType}
\alias{checkScalar}
\alias{checkLength}
\alias{checkLengths}
\alias{checkDim}
\alias{checkNames}
\alias{checkColnames}
\alias{checkMissing}
\alias{checkInfinite}
\alias{checkNumbers}
\alias{checkConstant}
\alias{checkMatrix}
\title{Check data inputs}
\usage{
checkType(x, expected)

checkScalar(x, expected)

checkLength(x, expected)

checkLengths(x, expected)

checkDim(x, expected)

checkNames(x, expected)

checkColnames(x, expected)

checkMissing(x)

checkInfinite(x)

checkNumbers(x, expected, ...)

checkConstant(x, expected, na.rm = TRUE)

checkMatrix(x, expected)
}
\arguments{
\item{x}{An object to be checked.}

\item{expected}{An appropriate expected value.}
}
\value{
Throw an error if any.
}
\description{
Check data inputs
}
\examples{
\dontrun{
object <- 1:3
checkType(object, "character")
checkScalar(object, "numeric")
checkLength(object, 2)
checkNames(object)
checkNames(object, c("a", "b", "c"))

object <- c(1:3, NA, Inf)
checkMissing(object)
checkInfinite(object)

object <- list(1:3, NULL, LETTERS)
checkLength(object, 2)
checkLengths(object)
checkLengths(object, c(3, 1, 26))

object <- matrix(-0.1, nrow = 3, ncol = 2)
checkDim(object, c(2, 3))
checkMatrix(object, "square")
checkMatrix(object, "symmetric")
checkNumbers(object, "positive", strict = TRUE)
checkNumbers(object, "whole")
}
}
\author{
N. Frerebeau
}
\keyword{internal}
