% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{combination}
\alias{combination}
\title{Binomial coefficient}
\usage{
combination(n, k)
}
\arguments{
\item{n}{A length-one \code{\link{numeric}} vector.}

\item{k}{A length-one \code{\link{numeric}} vector.}
}
\value{
A length-one \code{\link{numeric}} vector.
}
\description{
Computes the number of \code{k}-combinations from a given set of \code{n}
 elements ("\code{n} choose \code{k}").
}
\details{
Ramanujan approximation is used for \eqn{x!} computation if \eqn{x > 170}.
}
\references{
Ramanujan Aiyangar, S. (1988). \emph{The lost notebook and other unpublished
 papers}. Berlin: Springer-Verlag.
}
\author{
N. Frerebeau
}
\keyword{internal}
