% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compiegne}
\alias{compiegne}
\title{Compiègne ceramics}
\format{A numeric matrix with 5 rows (chronological periods, numbered from
 the oldest to the most recent from 1 to 5) and 16 variables (ceramic types):
 \describe{
  \item{A}{Red to white ceramics with fine sized inclusions.}
  \item{B}{Red to white ceramics with medium sized inclusions.}
  \item{C}{Dark ceramics with fine sized inclusions.}
  \item{D}{Dark ceramics with medium sized inclusions.}
  \item{E}{Ceramics close to those of groups B or D, with similarities to
  group F.}
  \item{F}{Black, red or beige ceramics with coarse inclusions.}
  \item{G}{Red polished ceramics with fine to medium sized inclusions.}
  \item{H}{Black polished ceramics with fine sized inclusions.}
  \item{I}{Black polished ceramics with medium sized inclusions.}
  \item{J}{Polished and painted ceramics with fine to medium sized
  inclusions.}
  \item{K}{Painted ceramics, similar to those of group A.}
  \item{L}{Painted ceramics, similar to those of group B.}
  \item{M}{Painted ceramics with coarse inclusions.}
  \item{N}{Glazed ceramics.}
  \item{O}{Stamped ceramics.}
  \item{P}{Coated ceramics.}
 }}
\source{
Lacroix, M. C. (1997). La céramique médiévale du site des Hallettes à
 Compiègne (Oise). \emph{Revue archéologique de Picardie}. Numéro spécial,
 13(1), 135-168.
 DOI: \href{https://doi.org/10.3406/pica.1997.1945}{10.3406/pica.1997.1945}
}
\usage{
compiegne
}
\description{
A dataset containing the ceramic counts from the Place des Hallettes in
 Compiègne (Oise, France). The data are grouped into five periods of about a
 century, ranging from the 9th to the 14th century, and sixteen ceramic
 types.
}
\seealso{
Other datasets: \code{\link{boves}},
  \code{\link{merzbach}}, \code{\link{mississippi}},
  \code{\link{zuni}}
}
\concept{datasets}
\keyword{datasets}
