% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CompositionIndex-class}
\alias{CompositionIndex-class}
\alias{.CompositionIndex}
\title{Composition Index}
\description{
An S4 class to represent an composition measure.
}
\note{
This class inherits from \linkS4class{DiversityIndex}.
}
\seealso{
\code{\link[=composition]{composition()}}

Other classes: 
\code{\link{DiversityIndex-class}},
\code{\link{EvennessIndex-class}},
\code{\link{HeterogeneityIndex-class}},
\code{\link{RarefactionIndex-class}},
\code{\link{RichnessIndex-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
