% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{conditions}
\alias{conditions}
\alias{catchConditions}
\alias{throwError}
\alias{formatErrors}
\title{Conditions}
\usage{
catchConditions(expr)

throwError(arg, must, not = NULL, call = NULL, ...)

formatErrors(object, errors)
}
\arguments{
\item{expr}{An expression.}

\item{arg}{A \code{\link{character}} string specifying the argument name.}

\item{must, not}{A \code{\link{character}} string specifying the error
message.}

\item{call}{The call.}

\item{...}{Extra arguments.}

\item{object}{An object to which error messages are related.}

\item{errors}{A \code{\link{character}} vector giving the error messages.}
}
\value{
Throw an error if \code{errors} is of non-zero length, returns \code{TRUE}
 if not.
}
\description{
Conditions
}
\author{
N. Frerebeau
}
\keyword{internal}
