% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{confidence}
\alias{confidence}
\title{Confidence interval for a proportion}
\usage{
confidence(x, level = 0.95, type = c("normal", "student"))
}
\arguments{
\item{x}{A \code{\link{numeric}} vector.}

\item{level}{A length-one \code{\link{numeric}} vector giving the significance
level to be used.}

\item{type}{A \code{\link{character}} string giving the type of confidence
interval to be returned. It must be one \code{normal} (default) or
\code{student}. Any unambiguous substring can be given.}
}
\value{
A \code{\link{numeric}} vector giving the margin of errors.
}
\description{
Computes the margin of errors of a confidence interval at a desired level of
 significance.
}
\author{
N. Frerebeau
}
\keyword{internal}
