% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{data.frame}
\alias{data.frame}
\alias{as.data.frame,DiversityIndex-method}
\title{Coerce to a Data Frame}
\usage{
\S4method{as.data.frame}{DiversityIndex}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{row.names, optional}{Currently not used.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Coerce to a Data Frame
}
\seealso{
Other mutators: 
\code{\link{mutators}}
}
\author{
N. Frerebeau
}
\concept{mutators}
