% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dating.R
\docType{methods}
\name{dateEvent}
\alias{dateEvent}
\alias{dateEvent-method}
\alias{dateAccumulation}
\alias{dateAccumulation-method}
\alias{dateEvent,CountMatrix-method}
\title{Date archaeological contexts}
\usage{
dateEvent(object, ...)

dateAccumulation(object, ...)

\S4method{dateEvent}{CountMatrix}(object, time, axes = 10,
  level = 0.95, verbose = FALSE, ...)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} matrix of count data.}

\item{...}{Further arguments to be passed to \code{\link[FactoMineR]{CA}}.}

\item{time}{A named \code{\link{numeric}} vector giving the reliable known
dates (in years) of the archaeological contexts.
Contexts are matched by names.}

\item{axes}{An \code{\link{integer}} giving the number of dimensions kept in
the results of the Correspondance Analysis (by default 10).}

\item{level}{A length-one \code{\link{numeric}} vector giving the
confidence level.}
}
\value{
An object of class \linkS4class{DateEvent} or \linkS4class{DateAccumulation}.
}
\description{
Date archaeological contexts
}
\references{
Bellanger, L. & Husi, P. (2012). Statistical Tool for Dating and
 Interpreting Archaeological Contexts Using Pottery. \emph{Journal of
 Archaeological Science}, 39(4), 777-790.
 DOI: \href{https://doi.org/10.1016/j.jas.2011.06.031}{10.1016/j.jas.2011.06.031}.
}
\author{
N. Frerebeau
}
