% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DateAccumulation-class}
\alias{DateAccumulation-class}
\title{Accumulation time}
\description{
An S4 class to represent the accumulation time of an archaeological
 assemblage.
}
\section{Slots}{

\describe{
\item{\code{assemblage}}{A \code{\link{character}} vector giving the name of the
archaeological assemblage.}

\item{\code{level}}{A length-one \code{\link{numeric}} vector giving the
confidence level.}

\item{\code{date}}{An \code{\link{integer}} vector giving the modelled date (in
years, calendar time) of the accumulation.}

\item{\code{earliest}}{An \code{\link{integer}} vector giving the modelled earliest
possible date (in years, calendar time) of the accumulation.}

\item{\code{latest}}{An \code{\link{integer}} vector giving the modelled latest
possible date (in years, calendar time) of the accumulation.}
}}

\author{
N. Frerebeau
}
