% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DateEvent-class}
\alias{DateEvent-class}
\alias{[[,DateEvent-method}
\alias{rows,DateEvent-method}
\alias{columns,DateEvent-method}
\title{Event time}
\usage{
\S4method{[[}{DateEvent}(x, i)

\S4method{rows}{DateEvent}(x)

\S4method{columns}{DateEvent}(x)
}
\arguments{
\item{x}{A \code{DateEvent} object from which to extract element(s).}

\item{i}{A \code{\link{character}} string specifying elements to extract.
Character vectors will be matched to the name of the slots.}
}
\description{
An S4 class to represent the event time of an archaeological assemblage.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[[}: Returns information about the individual slots.

\item \code{rows}: Returns the predicted dates for each archaeological
assemblage, the corresponding confidence interval and standard error of
the predicted dates.

\item \code{columns}: Returns the predicted dates for each archaeological
type or fabric, the corresponding confidence interval and standard error of
the predicted dates.
}}

\section{Slots}{

\describe{
\item{\code{model}}{A \code{\link[stats:lm]{multiple linear model}} TODO.}

\item{\code{level}}{A length-one \code{\link{numeric}} vector giving the
confidence level.}

\item{\code{residual}}{A length-one \code{\link{numeric}} vector giving the residual
standard deviation.}

\item{\code{rows}}{A numeric \code{\link{matrix}} giving the TODO.}

\item{\code{columns}}{A numeric \code{\link{matrix}} giving the TODO.}
}}

\author{
N. Frerebeau
}
