% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DateMCD-class}
\alias{DateMCD-class}
\alias{.DateMCD}
\title{Mean Ceramic Date}
\description{
S4 classes to store the Mean Ceramic Date of archaeological assemblages.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{\link{numeric}} matrix of count data.}

\item{\code{dates}}{A \code{\link{numeric}} vector of dates.}

\item{\code{mcd_values}}{A \code{\link{numeric}} vector of errors.}

\item{\code{mcd_errors}}{A \code{\link{numeric}} vector of errors.}
}}

\section{Subset}{

 In the code snippets below, \code{x} is a \code{DateMCD} object.
 \describe{
  \item{\code{x[[i]]}}{Extracts informations from a slot selected by
  subscript \code{i}. \code{i} is a length-one \code{\link{character}}
  vector. Returns the corresponding slot values.}
 }
}

\seealso{
Other class: 
\code{\link{DateEvent-class}},
\code{\link{DiversityIndex}},
\code{\link{IncrementTest-class}},
\code{\link{PermutationOrder-class}},
\code{\link{RefineCA-class}}
}
\author{
N. Frerebeau
}
\concept{class}
