% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/access.R
\docType{class}
\name{DateModel-class}
\alias{DateModel-class}
\alias{[[,DateModel-method}
\alias{rows,DateModel-method}
\alias{columns,DateModel-method}
\alias{accumulation,DateModel-method}
\alias{method,DateModel-method}
\title{Date model}
\usage{
\S4method{[[}{DateModel}(x, i)

\S4method{rows}{DateModel}(x)

\S4method{columns}{DateModel}(x)

\S4method{accumulation}{DateModel}(x)
}
\arguments{
\item{x}{A \code{DateModel} object from which to extract element(s).}

\item{i}{A \code{\link{character}} string specifying elements to extract.
Character string will be matched to the name of the slots.}
}
\description{
An S4 class to store the event and accumulation times of archaeological
 assemblages.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[[}: Returns information about the individual slots.

\item \code{rows}: Returns a numeric matrix giving the predicted event
dates for each archaeological assemblage, the corresponding confidence
interval and standard error of the predicted dates.

\item \code{columns}: Returns a numeric matrix giving the predicted event
dates for each archaeological type or fabric, the corresponding confidence
interval and standard error of the predicted dates.

\item \code{accumulation}: Returns the accumulation date for each archaeological
assemblage.
}}

\section{Slots}{

\describe{
\item{\code{counts}}{A \eqn{m \times p}{m x p} \code{\link{numeric}} matrix of count
data used for model fitting.}

\item{\code{dates}}{A two columns \code{\link{data.frame}} giving the known dates
used for model fitting and an identifier to link each row to an assemblage.}

\item{\code{level}}{A length-one \code{\link{numeric}} vector giving the
confidence level.}

\item{\code{model}}{A \code{\link[stats:lm]{multiple linear model}}: the Gaussian
multiple linear regression model fitted for event date estimation and
prediction.}

\item{\code{residual}}{A length-one \code{\link{numeric}} vector giving the residual
standard deviation.}

\item{\code{rows}}{A five columns \code{\link{data.frame}} giving the predicted event
dates for each archaeological assemblage, the corresponding confidence
interval boundaries, standard error of the predicted dates and an
identifier to link each row to an assemblage.}

\item{\code{columns}}{A five columns \code{\link{data.frame}} giving the predicted
event dates for each archaeological type or fabric, the corresponding
confidence interval boundaries, standard error of the predicted dates and an
identifier to link each row to a type or fabric.}

\item{\code{accumulation}}{A two columns \code{\link{data.frame}} giving the point
estimate of accumulation dates of archaeological assemblages and an
identifier to link each row to an assemblage.}
}}

\author{
N. Frerebeau
}
