% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecate.R
\docType{methods}
\name{tabula-deprecated}
\alias{tabula-deprecated}
\alias{plotBar}
\alias{plotMatrix}
\alias{plotRank}
\alias{plotSpot}
\alias{seriate}
\alias{refine}
\alias{diversity}
\alias{evenness}
\alias{richness}
\alias{refine,AbsoluteFrequencyMatrix-method}
\alias{refine,DateModel-method}
\alias{plotBar,AbsoluteFrequencyMatrix-method}
\alias{plotBar,RelativeFrequencyMatrix-method}
\alias{plotMatrix,AbsoluteFrequencyMatrix-method}
\alias{plotMatrix,RelativeFrequencyMatrix-method}
\alias{plotMatrix,IncidenceMatrix-method}
\alias{plotRank,AbsoluteFrequencyMatrix-method}
\alias{plotRank,RelativeFrequencyMatrix-method}
\alias{plotSpot,AbsoluteFrequencyMatrix-method}
\alias{plotSpot,RelativeFrequencyMatrix-method}
\alias{plotSpot,SimilarityMatrix-method}
\alias{plotSpot,OccurrenceMatrix-method}
\alias{seriate,AbsoluteFrequencyMatrix,missing-method}
\alias{seriate,IncidenceMatrix,missing-method}
\alias{seriate,AbsoluteFrequencyMatrix,BootCA-method}
\alias{diversity,AbsoluteFrequencyMatrix-method}
\alias{evenness,AbsoluteFrequencyMatrix-method}
\alias{richness,AbsoluteFrequencyMatrix-method}
\alias{richness,IncidenceMatrix-method}
\title{Deprecated Methods}
\usage{
plotBar(object, ...)

plotMatrix(object, ...)

plotRank(object, ...)

plotSpot(object, ...)

seriate(object, subset, ...)

refine(object, ...)

diversity(object, ...)

evenness(object, ...)

richness(object, ...)

\S4method{refine}{AbsoluteFrequencyMatrix}(object, cutoff, n = 1000,
  axes = c(1, 2), ...)

\S4method{refine}{DateModel}(object, method = c("jackknife",
  "bootstrap"), n = 1000, ...)

\S4method{plotBar}{AbsoluteFrequencyMatrix}(object, level = FALSE,
  EPPM = FALSE, center = TRUE, horizontal = FALSE)

\S4method{plotBar}{RelativeFrequencyMatrix}(object, level = FALSE,
  EPPM = FALSE, center = TRUE, horizontal = FALSE)

\S4method{plotMatrix}{AbsoluteFrequencyMatrix}(object, PVI = FALSE)

\S4method{plotMatrix}{RelativeFrequencyMatrix}(object, PVI = FALSE)

\S4method{plotMatrix}{IncidenceMatrix}(object)

\S4method{plotRank}{AbsoluteFrequencyMatrix}(object, PVI = FALSE)

\S4method{plotRank}{RelativeFrequencyMatrix}(object, PVI = FALSE)

\S4method{plotSpot}{AbsoluteFrequencyMatrix}(object, threshold = NULL)

\S4method{plotSpot}{RelativeFrequencyMatrix}(object, threshold = NULL)

\S4method{plotSpot}{SimilarityMatrix}(object)

\S4method{plotSpot}{OccurrenceMatrix}(object)

\S4method{seriate}{AbsoluteFrequencyMatrix,missing}(object,
  method = c("correspondance", "reciprocal"), EPPM = FALSE,
  margin = c(1, 2), stop = 100, ...)

\S4method{seriate}{IncidenceMatrix,missing}(object,
  method = c("correspondance", "reciprocal"), margin = c(1, 2),
  stop = 100, ...)

\S4method{seriate}{AbsoluteFrequencyMatrix,BootCA}(object, subset,
  margin = c(1, 2), ...)

\S4method{diversity}{AbsoluteFrequencyMatrix}(object,
  method = c("berger", "brillouin", "mcintosh", "shannon", "simpson"),
  simplify = FALSE, ...)

\S4method{evenness}{AbsoluteFrequencyMatrix}(object,
  method = c("brillouin", "mcintosh", "shannon", "simpson"),
  simplify = FALSE, ...)

\S4method{richness}{AbsoluteFrequencyMatrix}(object, method = c("ace",
  "chao1", "margalef", "menhinick", "none"), unbiased = FALSE,
  improved = FALSE, k = 10, simplify = FALSE)

\S4method{richness}{IncidenceMatrix}(object, method = c("chao2", "ice"),
  unbiased = FALSE, improved = FALSE, k = 10, simplify = FALSE)
}
\arguments{
\item{object}{An object.}

\item{...}{Further arguments to be passed to internal methods.}

\item{subset}{A \linkS4class{BootCA} object giving the subset of
\code{object} to be used.}

\item{cutoff}{A function that takes a numeric vector as argument and returns
a single numeric value.}

\item{n}{A non-negative \code{\link{integer}} giving the number of partial
bootstrap replications.}

\item{axes}{A \code{\link{numeric}} vector giving the subscripts of the CA
axes to use.}

\item{method}{A \code{\link{character}} string specifying the method to be
used. This must be one of "\code{reciprocal}", "\code{correspondence}".
Any unambiguous substring can be given.}

\item{level}{A length-one \code{\link{numeric}} vector giving the
confidence level to be drawn.}

\item{EPPM}{A \code{\link{logical}} scalar: should the EPPM be drawn (see
details)?}

\item{center}{A \code{\link{logical}} scalar: should the bar plot
be centered? The default, \code{TRUE}, produces a Ford diagram, otherwise it
produces a Bertin diagram.}

\item{horizontal}{A \code{\link{logical}} scalar: should the bar plot
be horizontal? The default, \code{FALSE}, means variables in rows and
cases in columns (i.e. Bertin diagram). Only used if \code{center} is
\code{FALSE}.}

\item{PVI}{A \code{\link{logical}} scalar: should the PVI be drawn instead of
frequencies (see details)?}

\item{margin}{A \code{\link{numeric}} vector giving the subscripts which the
rearrangement will be applied over: \code{1} indicates rows, \code{2}
indicates columns, \code{c(1, 2)} indicates rows then columns,
\code{c(2, 1)} indicates columns then rows.}

\item{stop}{A length-one \code{\link{numeric}} vector giving the stopping rule
(i.e. maximum number of iterations) to avoid infinite loop.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? The default value, \code{FALSE}, returns a list.}

\item{unbiased}{A \code{\link{logical}} scalar. Should the bias-corrected
estimator be used? Only used with "\code{chao1}" or "\code{chao2}"
(improved) estimator.}

\item{improved}{A \code{\link{logical}} scalar. Should the improved
estimator be used? Only used with "\code{chao1}" or "\code{chao2}".}

\item{k}{A length-one \code{\link{numeric}} vector giving the threshold
between rare/infrequent and abundant/frequent species. Only used if
\code{method} is "\code{ace}" or "\code{ice}".}

\item{log}{A \code{\link{character}} string which contains "\code{x}" if the
x axis is to be logarithmic, "\code{y}" if the y axis is to be logarithmic
and "\code{xy}" or "\code{yx}" if both axes are to be logarithmic (base 10).}

\item{facet}{A \code{\link{logical}} scalar: should a matrix of panels
defined by case/sample be drawn?}

\item{EPPM}{A \code{\link{logical}} scalar: should the seriation be computed
on EPPM instead of raw data?}

\item{sample}{A length-one \code{\link{numeric}} vector giving the sub-sample
size.}
}
\description{
\code{plotBar} produces a Bertin or a Ford (battleship curve) diagram.
}
\keyword{internal}
