% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EvennessIndex-class}
\alias{EvennessIndex-class}
\alias{.EvennessIndex}
\title{Evenness Index}
\description{
An S4 class to represent an evenness measure.
}
\note{
This class inherits from \linkS4class{DiversityIndex}.
}
\seealso{
\code{\link[=evenness]{evenness()}}

Other classes: 
\code{\link{CompositionIndex-class}},
\code{\link{DiversityIndex-class}},
\code{\link{HeterogeneityIndex-class}},
\code{\link{RarefactionIndex-class}},
\code{\link{RichnessIndex-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
