% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{FIT}
\alias{FIT}
\title{Frequency Increment Test (FIT)}
\usage{
FIT(v, t, ...)
}
\arguments{
\item{v}{A \code{\link{numeric}} vector of frequencies.}

\item{t}{A \code{\link{numeric}} vector of time coordinates.}

\item{...}{Extra parameter passed to \code{\link[=stats]{t.test}}.}
}
\value{
A \code{\link{numeric}} vector containing the following components:
 \describe{
  \item{t}{the value of the test statistic.}
  \item{p.value}{the p-value for the test.}
 }
}
\description{
Frequency Increment Test (FIT)
}
\references{
Feder, A. F., Kryazhimskiy, S. & Plotkin, J. B. (2014). Identifying
 Signatures of Selection in Genetic Time Series. \emph{Genetics}, 196(2),
 509-522.
 DOI: \href{https://doi.org/10.1534/genetics.113.158220}{10.1534/genetics.113.158220}.
}
\author{
N. Frerebeau
}
\keyword{internal}
