% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generateUUID}
\alias{generateUUID}
\title{UUID v4}
\usage{
generateUUID(seed = NULL)
}
\arguments{
\item{seed}{A single \code{\link{integer}} specifying the seeds.
If \code{NULL} (the default) the seed will be re-initialized.}
}
\value{
A 36 characters long \code{\link{character}} string.
}
\description{
Generates a universally unique identifier (UUID v4).
}
\details{
As it rely on R's internal random number generators and so will suffer
 from the use of \code{\link{set.seed}} in a session, the seed is
 re-initialized during execution (unless \code{seed} is not \code{NULL}).
 To prevent any side effects, the random number generator (RNG) state is
 saved and restored when the function exits.
}
\seealso{
\link{set.seed}
}
\author{
N. Frerebeau
}
\keyword{internal}
