% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{helpers}
\alias{helpers}
\alias{\%||\%}
\alias{\%o\%}
\alias{compact}
\alias{detect}
\alias{count}
\title{Helpers}
\usage{
lhs \%||\% rhs

f \%o\% g

compact(f, x)

detect(f, x)

count(f, x)
}
\arguments{
\item{lhs}{An object.}

\item{rhs}{An object.}

\item{f, f1, f2}{A \code{\link{function}}. In \code{compact} and \code{detect}
\code{f} must be a logical predicate.}

\item{g}{A \code{\link{function}}.}

\item{x}{An object.}
}
\value{
TODO
}
\description{
\code{compact} removes elements from a list or vector.
\code{detect} xxx.
}
\details{
%o% allows for function composition.

%||% xxx.


 Adapted from H. Wickham's \emph{Avanced R}.
}
\references{
Wickham, H. (2014). \emph{Advanced R}. London: Chapman & Hall. The R Series.
}
\keyword{internal}
