% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HeterogeneityIndex-class}
\alias{HeterogeneityIndex-class}
\alias{.HeterogeneityIndex}
\title{Heterogeneity Index}
\description{
An S4 class to represent an heterogeneity measure.
}
\note{
This class inherits from \linkS4class{DiversityIndex}.
}
\seealso{
\code{\link[=heterogeneity]{heterogeneity()}}

Other classes: 
\code{\link{CompositionIndex-class}},
\code{\link{DiversityIndex-class}},
\code{\link{EvennessIndex-class}},
\code{\link{RarefactionIndex-class}},
\code{\link{RichnessIndex-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
