% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IncrementTest-class}
\alias{IncrementTest-class}
\alias{.IncrementTest}
\title{Frequency Increment Test}
\description{
An S4 class to represent a Frequency Increment Test results.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{\link{numeric}} matrix of count data.}

\item{\code{dates}}{A \code{\link{numeric}} vector of dates.}

\item{\code{statistic}}{A \code{\link{numeric}} vector giving the values of the
t-statistic.}

\item{\code{parameter}}{An \code{\link{integer}} giving the degrees of freedom for
the t-statistic.}

\item{\code{p_value}}{A \code{\link{numeric}} vector giving the the p-value for the
test.}
}}

\section{Subset}{

 In the code snippets below, \code{x} is a \code{IncrementTest} object.
 \describe{
  \item{\code{x[[i]]}}{Extracts informations from a slot selected by
  subscript \code{i}. \code{i} is a length-one \code{\link{character}}
  vector. Returns the corresponding slot values.}
 }
}

\seealso{
Other class: 
\code{\link{DateEvent-class}},
\code{\link{DateMCD-class}},
\code{\link{DiversityIndex}},
\code{\link{PermutationOrder-class}},
\code{\link{RefineCA-class}}
}
\author{
N. Frerebeau
}
\concept{class}
