% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{independance}
\alias{independance}
\title{Independance}
\usage{
independance(x, method = c("EPPM", "PVI"))
}
\arguments{
\item{x}{A \eqn{m \times p}{m x p} \link{\code{numeric}} matrix.}

\item{method}{A \link{\code{character}} string giving the method to be used.
This must be one of "\code{EPPM}" or "\code{PVI}" (see details). Any
unambiguous substring can be given.}
}
\value{
A \eqn{m \times p}{m x p} \link{\code{numeric}} matrix.
}
\description{
Independance
}
\details{
Computes for each cell of a numeric matrix one of the following thresholds:
 \describe{
  \item{EPPM}{The positive deviation from the column mean percentage (in
   french "écart positif au pourcentage moyen", EPPM)}
  \item{PVI}{The percentage of independence value (in french,
  "pourcentage de valeur d'indépendance", PVI).}
 }
}
\author{
N. Frerebeau
}
\keyword{internal}
