% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-diversity.R
\name{index-heterogeneity}
\alias{index-heterogeneity}
\alias{dominanceBerger}
\alias{diversityBrillouin}
\alias{evennessBrillouin}
\alias{dominanceMcintosh}
\alias{evennessMcintosh}
\alias{diversityShannon}
\alias{evennessShannon}
\alias{varianceShannon}
\alias{dominanceSimpson}
\alias{evennessSimpson}
\title{Diversity, dominance and evenness index}
\usage{
dominanceBerger(x, ...)

diversityBrillouin(x, ...)

evennessBrillouin(x, ...)

dominanceMcintosh(x, ...)

evennessMcintosh(x, ...)

diversityShannon(x, base = exp(1), ...)

evennessShannon(x, ...)

varianceShannon(x, ...)

dominanceSimpson(x, ...)

evennessSimpson(x, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector giving the number of individuals for
each class.}

\item{...}{Currently not used.}
}
\value{
A length-one \code{\link{numeric}} vector.
}
\description{
\code{dominanceBerger} returns Berger-Parker dominance index.
\code{diversityBrillouin} and \code{evennessBrillouin} return Brillouin
diversity index and evenness.
\code{dominanceMcintosh} and \code{evennessMcintosh} return McIntosh
dominance index and evenness.
\code{diversityShannon}, \code{evennessShannon}, \code{varianceShannon}
return Shannon-Wiener diversity index, evenness and variance.
\code{dominanceSimpson} and \code{evennessSimpson} return Simpson dominance
index and evenness.
}
\references{
Berger, W. H. & Parker, F. L. (1970). Diversity of Planktonic Foraminifera
 in Deep-Sea Sediments. \emph{Science}, 168(3937), 1345-1347.
 DOI: \href{https://doi.org/10.1126/science.168.3937.1345}{10.1126/science.168.3937.1345}.

 Brillouin, L. (1956). \emph{Science and information theory}. New York:
 Academic Press.

 McIntosh, R. P. (1967). An Index of Diversity and the Relation of Certain
 Concepts to Diversity. \emph{Ecology}, 48(3), 392-404.
 DOI: \href{https://doi.org/10.2307/1932674}{10.2307/1932674}.

 Shannon, C. E. (1948). A Mathematical Theory of Communication. \emph{The
 Bell System Technical Journal}, 27, 379-423.
 DOI: \href{https://doi.org/10.1002/j.1538-7305.1948.tb01338.x}{10.1002/j.1538-7305.1948.tb01338.x}.

 Simpson, E. H. (1949). Measurement of Diversity. \emph{Nature}, 163(4148),
 688-688. DOI: \href{https://doi.org/10.1038/163688a0}{10.1038/163688a0}.
}
\author{
N. Frerebeau
}
\concept{diversity index}
\keyword{internal}
