% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-rarefaction.R
\name{index-rarefaction}
\alias{index-rarefaction}
\alias{rarefactionHurlbert}
\title{Rarefaction index}
\usage{
rarefactionHurlbert(x, sample)
}
\arguments{
\item{n}{A \code{\link{numeric}} vector giving the number of individuals for
each type.}

\item{...}{Currently not used.}
}
\value{
A length-one \code{\link{numeric}} vector.
}
\description{
\code{rarefactionHurlbert} returns Hurlbert's unbiaised estimate of Sander's
rarefaction.
}
\references{
Hurlbert, S. H. (1971). The Nonconcept of Species Diversity: A Critique and
 Alternative Parameters. \emph{Ecology}, 52(4), 577-586.
 DOI: \href{https://doi.org/10.2307/1934145}{10.2307/1934145}.

 Sander, H. L. (1968). Marine Benthic Diversity: A Comparative Study.
 \emph{The American Naturalist}, 102(925), 243-282.
}
\seealso{
Other diversity measures: \code{\link{index-richness}},
  \code{\link{index-similarity}},
  \code{\link{index-turnover}}
}
\author{
N. Frerebeau
}
\concept{diversity measures}
\keyword{internal}
