% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-similarity.R
\name{index-similarity}
\alias{index-similarity}
\alias{similarityJaccard}
\alias{similaritySorenson}
\alias{similarityBray}
\alias{similarityMorisita}
\alias{similarityBrainerd}
\alias{similarityBinomial}
\title{Similarity index}
\usage{
similarityJaccard(x, y)

similaritySorenson(x, y)

similarityBray(x, y)

similarityMorisita(x, y)

similarityBrainerd(x, y)

similarityBinomial(x, y)
}
\arguments{
\item{x}{A length-p \code{\link{numeric}} vector.}

\item{y}{A length-p \code{\link{numeric}} vector.}
}
\value{
A length-one \code{\link{numeric}} vector.
}
\description{
Qualitative index:
\code{similarityJaccard} returns Jaccard similarity index.
\code{similaritySorenson} returns Sorenson similarity index.

Quantitative index:
\code{similarityBray} returns Bray and Curtis modified version of the
Sorenson index.
\code{similarityMorisita} returns Morisita-Horn quantitative index.
\code{similarityBrainerd} returns Brainerd-Robinson quantitative index.
\code{similarityBinomial} retunrs binomial co-occurrence of types assessment.
}
\seealso{
Other diversity measures: \code{\link{index-rarefaction}},
  \code{\link{index-richness}},
  \code{\link{index-turnover}}
}
\author{
N. Frerebeau
}
\concept{diversity measures}
\keyword{internal}
