% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-turnover.R
\name{index-turnover}
\alias{index-turnover}
\alias{turnoverWhittaker}
\alias{turnoverCody}
\alias{turnoverRoutledge1}
\alias{turnoverRoutledge2}
\alias{turnoverRoutledge3}
\alias{turnoverWilson}
\title{Turnover index}
\usage{
turnoverWhittaker(x)

turnoverCody(x)

turnoverRoutledge1(x)

turnoverRoutledge2(x)

turnoverRoutledge3(x)

turnoverWilson(x)
}
\arguments{
\item{x}{A \code{\link{logical}} or \code{\link{numeric}} matrix.}
}
\value{
A length-one \code{\link{numeric}} vector.
}
\description{
Turnover index
}
\seealso{
Other diversity measures: \code{\link{index-rarefaction}},
  \code{\link{index-richness}},
  \code{\link{index-similarity}}
}
\author{
N. Frerebeau
}
\concept{diversity measures}
\keyword{internal}
