% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity.R
\docType{methods}
\name{index_bray}
\alias{index_bray}
\alias{index_bray-method}
\alias{index_bray,numeric,numeric-method}
\title{Sorenson Quantitative Index}
\usage{
index_bray(x, y, ...)

\S4method{index_bray}{numeric,numeric}(x, y)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Bray and Curtis modified version of the Sorenson index.
}
\references{
Bray, J. R. & Curtis, J. T. (1957). An Ordination of the Upland Forest
Communities of Southern Wisconsin. \emph{Ecological Monographs}, 27(4),
325-349. \doi{10.2307/1942268}.
}
\seealso{
Other beta diversity measures: 
\code{\link{index_binomial}()},
\code{\link{index_brainerd}()},
\code{\link{index_cody}()},
\code{\link{index_jaccard}()},
\code{\link{index_morisita}()},
\code{\link{index_routledge}},
\code{\link{index_sorenson}()},
\code{\link{index_whittaker}()},
\code{\link{index_wilson}()}
}
\author{
N. Frerebeau
}
\concept{beta diversity measures}
