% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/similarity.R
\docType{methods}
\name{index_jaccard}
\alias{index_jaccard}
\alias{index_jaccard-method}
\alias{index_jaccard,character,character-method}
\alias{index_jaccard,logical,logical-method}
\alias{index_jaccard,numeric,numeric-method}
\title{Jaccard Index}
\usage{
index_jaccard(x, y, ...)

\S4method{index_jaccard}{character,character}(x, y)

\S4method{index_jaccard}{logical,logical}(x, y)

\S4method{index_jaccard}{numeric,numeric}(x, y)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Jaccard Index
}
\details{
Data are standardized on a presence/absence scale (\eqn{0}/\eqn{1})
beforehand.
}
\references{
Magurran, A. E. (1988). \emph{Ecological Diversity and its Measurement}.
Princeton, NJ: Princeton University Press. \doi{10.1007/978-94-015-7358-0}.
}
\seealso{
Other beta diversity measures: 
\code{\link{index_binomial}()},
\code{\link{index_brainerd}()},
\code{\link{index_bray}()},
\code{\link{index_cody}()},
\code{\link{index_morisita}()},
\code{\link{index_routledge}},
\code{\link{index_sorensen}()},
\code{\link{index_whittaker}()},
\code{\link{index_wilson}()}
}
\author{
N. Frerebeau
}
\concept{beta diversity measures}
