% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{jackknife}
\alias{jackknife}
\title{Jackknife estimation}
\usage{
jackknife(x, do, ...)
}
\arguments{
\item{x}{A vector.}

\item{do}{A \code{\link{function}} that takes \code{x} as an argument
and returns a single numeric value.}

\item{...}{Extra arguments passed to \code{do}.}
}
\value{
A list with the following elements:
 \describe{
  \item{values}{The \eqn{n} leave-one-out values.}
  \item{bias}{The jackknife estimate of bias.}
  \item{error}{he jackknife estimate of standard error.}
 }
}
\description{
Jackknife estimation
}
\keyword{internal}
