% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LogicalMatrix-class}
\alias{LogicalMatrix-class}
\title{Logical matrix}
\description{
An S4 class to represent a logical matrix.
}
\seealso{
\linkS4class{Matrix}
}
\author{
N. Frerebeau
}
\keyword{internal}
