% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/LogicalMatrix.R
\docType{class}
\name{LogicalMatrix}
\alias{LogicalMatrix}
\alias{IncidenceMatrix}
\alias{StratigraphyMatrix}
\title{S4 classes to represent logical matrix}
\usage{
IncidenceMatrix(data = NA, nrow = 1, ncol = 1, byrow = FALSE,
  dimnames = NULL)

StratigraphyMatrix(data = NA, nrow = 1, ncol = 1, byrow = FALSE,
  dimnames = NULL)
}
\arguments{
\item{data}{an optional data vector (including a list or
    \code{\link{expression}} vector).  Non-atomic classed \R objects are
    coerced by \code{\link{as.vector}} and all attributes discarded.}

\item{nrow}{the desired number of rows.}

\item{ncol}{the desired number of columns.}

\item{byrow}{logical. If \code{FALSE} (the default) the matrix is
    filled by columns, otherwise the matrix is filled by rows.}

\item{dimnames}{A \code{\link{dimnames}} attribute for the matrix:
    \code{NULL} or a \code{list} of length 2 giving the row and column
    names respectively.  An empty list is treated as \code{NULL}, and a
    list of length one as row names.  The list can be named, and the
    list names will be used as names for the dimensions.}
}
\description{
S4 classes to represent logical matrix
}
\author{
N. Frerebeau
}
