% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Matrix-class}
\alias{Matrix-class}
\title{Matrix}
\description{
An S4 class to represent a matrix.
}
\section{Slots}{

\describe{
\item{\code{uuid}}{A \code{\link{character}} string specifying the unique
identifier of the object.}

\item{\code{cases}}{A \code{\link{character}} vector specifying the row names.}

\item{\code{types}}{A \code{\link{character}} vector specifying the column names.}
}}

\note{
This class extends the \code{base} \link[base]{matrix}.
}
\seealso{
\link[base]{matrix}
}
\author{
N. Frerebeau
}
\keyword{internal}
