% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merzbach}
\alias{merzbach}
\title{Merzbach ceramics}
\format{
A numeric matrix with 8 rows (phases, numbered from VII to XIV)
and 36 variables (pottery motifs).
 \describe{
  \item{BT1, BT2, ...}{ Counts of a motif.}
 }
}
\source{
Crema, E. R., Kandler, A. & Shennan, S. (2016). Revealing Patterns of
 Cultural Transmission from Frequency Data: Equilibrium and Non-Equilibrium
 Assumptions. \emph{Scientific Reports}, 6(1).
 DOI: \href{https://doi.org/10.1038/srep39122}{10.1038/srep39122}.

 Crema, E. R. (2016). Sample codes and data for "Revealing patterns of
 cultural transmission from frequency data: equilibrium and non-equilibrium
 assumptions". \emph{Zenodo}, v1.0.
 DOI: \href{https://doi.org/10.5281/zenodo.187558}{10.5281/zenodo.187558}.
}
\usage{
merzbach
}
\description{
A dataset containing the ceramic counts from the Merzbach assemblage
(Germany). The data are grouped into eight phases.
}
\seealso{
Other datasets: 
\code{\link{boves}},
\code{\link{compiegne}},
\code{\link{mississippi}},
\code{\link{zuni}}
}
\concept{datasets}
\keyword{datasets}
