% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mississippi}
\alias{mississippi}
\title{Mississippi ceramics}
\format{
A numeric matrix with 20 rows and 10 variables (ceramic types):
 \describe{
   \item{ParkinPunctate}{}
   \item{BartonKentMPI}{}
   \item{Painted}{}
   \item{FortuneNoded}{}
   \item{RanchIncised}{}
   \item{WallsEngraved}{}
   \item{WallaceIncised}{}
   \item{RhodesIncised}{}
   \item{VernonPaulApplique}{}
   \item{HullEngraved}{}
 }
}
\source{
Lipo, C. P., Madsen, M. E. & Dunnell, R. C. (2015). A
 Theoretically-Sufficient and Computationally-Practical Technique for
 Deterministic Frequency Seriation. \emph{PLOS ONE}, 10(4), e0124942.
 DOI: \href{https://doi.org/10.1371/journal.pone.0124942}{10.1371/journal.pone.0124942}.
}
\usage{
mississippi
}
\description{
A dataset containing ceramic counts from the Mississippi region.
}
\seealso{
Other datasets: 
\code{\link{boves}},
\code{\link{compiegne}},
\code{\link{merzbach}},
\code{\link{zuni}}
}
\concept{datasets}
\keyword{datasets}
