% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutator}
\alias{mutator}
\alias{get}
\alias{set}
\alias{get_index}
\alias{get_index-method}
\alias{get_method}
\alias{get_method-method}
\alias{get_order}
\alias{get_order-method}
\alias{get_index,HeterogeneityIndex-method}
\alias{get_index,EvennessIndex-method}
\alias{get_index,RichnessIndex-method}
\alias{get_method,DiversityIndex-method}
\title{Get or Set Parts of an Object}
\usage{
get_index(x)

get_method(x)

get_order(x)

\S4method{get_index}{HeterogeneityIndex}(x)

\S4method{get_index}{EvennessIndex}(x)

\S4method{get_index}{RichnessIndex}(x)

\S4method{get_method}{DiversityIndex}(x)
}
\arguments{
\item{x}{An object from which to get or set element(s).}
}
\value{
An object of the same sort as \code{object} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\seealso{
Other mutator: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutator}
