% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{get_method}
\alias{get_method-method}
\alias{get_method,DiversityIndex-method}
\title{Get or Set Parts of an Object}
\usage{
get_method(x)

\S4method{get_method}{DiversityIndex}(x)
}
\arguments{
\item{x}{An object from which to get or set element(s).}
}
\value{
An object of the same sort as \code{object} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\author{
N. Frerebeau
}
\concept{mutators}
