% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NumericMatrix-class}
\alias{NumericMatrix-class}
\title{Numeric matrix}
\description{
An S4 class to represent a numeric matrix.
}
\seealso{
\linkS4class{Matrix}
}
\author{
N. Frerebeau
}
\keyword{internal}
