% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diversity_alpha.R
\docType{methods}
\name{observed}
\alias{observed}
\alias{observed-method}
\alias{singleton}
\alias{singleton-method}
\alias{doubleton}
\alias{doubleton-method}
\alias{observed,numeric-method}
\alias{singleton,numeric-method}
\alias{doubleton,numeric-method}
\title{Number of Observed Species}
\usage{
observed(x, ...)

singleton(x, ...)

doubleton(x, ...)

\S4method{observed}{numeric}(x, na.rm = FALSE, ...)

\S4method{singleton}{numeric}(x, na.rm = FALSE, ...)

\S4method{doubleton}{numeric}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector of count data (absolute frequencies).}

\item{...}{Currently not used.}

\item{na.rm}{A \code{\link{numeric}} scalar: should missing values (including \code{NaN}) be
removed?}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Number of Observed Species
}
\seealso{
Other alpha diversity measures: 
\code{\link{index_ace}()},
\code{\link{index_baxter}()},
\code{\link{index_berger}()},
\code{\link{index_boone}()},
\code{\link{index_brillouin}()},
\code{\link{index_chao1}()},
\code{\link{index_chao2}()},
\code{\link{index_hurlbert}()},
\code{\link{index_ice}()},
\code{\link{index_margalef}()},
\code{\link{index_mcintosh}()},
\code{\link{index_menhinick}()},
\code{\link{index_shannon}()},
\code{\link{index_simpson}()},
\code{\link{index_squares}()}
}
\concept{alpha diversity measures}
