% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{OccurrenceMatrix-class}
\alias{OccurrenceMatrix-class}
\alias{.OccurrenceMatrix}
\title{Co-occurrence matrix}
\description{
An S4 class to represent a co-occurrence matrix.
}
\details{
A co-occurrence matrix is a symetric matrix with zeros on its main diagonal,
 which works out how many times (expressed in percent) each pairs of
 taxa/types occur together in at least one sample.
}
\section{Matrix ID}{

 When a matrix is first created, an identifier is generated (UUID v4).
 This ID is preserved when coercing to another class. Thus, the object ID is
 unique within the same class, but two objects of different classes can have
 the same ID. This makes it possible to identify objects representing the
 same initial data and associate them with the results of specific
 computations (e. g. \link[=seriate]{seriation}).
}

\section{Access}{

 In the code snippets below, \code{x} is a \code{*Matrix} object.
 \describe{
  \item{\code{dim(x)}}{Get the dimension of \code{x}.}
  \item{\code{colnames(x)}, \code{colnames(x) <- value}}{Get or set the
  column names of \code{x} according to \code{value}.}
  \item{\code{rownames(x)}, \code{rownames(x) <- value}}{Get or set the
  row names of \code{x} according to \code{value}.}
  \item{\code{getID(x)}}{Get the ID of \code{x}.}
 }
}

\section{Subset}{

 In the code snippets below, \code{x} is a \code{*Matrix} object.
 \describe{
  \item{\code{x[i, j]}}{Extracts elements selected by subscripts \code{i}
  and \code{j}. Indices are \code{\link{numeric}}, \code{\link{integer}} or
  \code{\link{character}} vectors or empty (missing) or \code{NULL}.
  Numeric values are coerced to \code{\link{integer}} as by
  \code{\link{as.integer}} (and hence truncated towards zero).
  Character vectors will be matched to the name of the elements.
  An empty index (a comma separated blank) indicates that all
  entries in that dimension are selected.
  Returns an object of the same class as \code{x}.}
  \item{\code{x[[i]]}}{Extracts informations from a slot selected by
  subscript \code{i}. \code{i} should be one of "\code{id}" or \code{NULL}.}
 }
}

\examples{
# Create a count data matrix
A1 <- CountMatrix(data = sample(0:10, 100, TRUE),
                  nrow = 10, ncol = 10, byrow = TRUE)

# Access
dim(A1) # Get the matrix dimensions
colnames(A1) # Get the column names
colnames(A1) <- letters[11:20] # Set the column names
rownames(A1) # Get the rownames
rownames(A1) <- LETTERS[1:10] # Set the row names

# Subset
A1[["id"]] # Get the matrix ID
A1[, ] # Get all values
A1[1, ] # Get the first row
A1[c("A", "B", "C"), ] # Get the first three rows
A1[c("A", "B", "C"), 1] # Get the first three rows of the first column
A1[, 1, drop = FALSE] # Get the first column

# Coerce counts to frequencies
B <- as(A1, "FrequencyMatrix")

# Row sums are internally stored before coercing to a frequency matrix
getTotals(B) # Get row sums

# This allows to restore the source data
A2 <- as(B, "CountMatrix")
all(A1 == A2)

# Coerce to a co-occurrence matrix
C <- as(B, "OccurrenceMatrix")
}
\seealso{
\linkS4class{NumericMatrix}

Other abundance matrix: \code{\link{CountMatrix-class}},
  \code{\link{FrequencyMatrix-class}}
}
\author{
N. Frerebeau
}
\concept{abundance matrix}
