% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PermutationOrder-class}
\alias{PermutationOrder-class}
\title{An S4 class to represent a permutation order}
\description{
An S4 class to represent a permutation order
}
\section{Slots}{

\describe{
\item{\code{rows}}{A \code{\link{integer}} vector.}

\item{\code{columns}}{A \code{\link{integer}} vector.}

\item{\code{seriation}}{A \code{\link{character}} vector.}
}}

\author{
N. Frerebeau
}
