% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{permute}
\alias{permute}
\title{Permute rows or columns}
\usage{
permute(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Further arguments passed to other methods.}
}
\description{
Permute rows or columns
}
\seealso{
\code{\link{seriate}}
}
\author{
N. Frerebeau
}
