% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot_diversity.R
\docType{methods}
\name{plot_diversity}
\alias{plot_diversity}
\alias{plot,DiversityIndex,missing-method}
\title{Diversity Plot}
\usage{
\S4method{plot}{DiversityIndex,missing}(x)
}
\arguments{
\item{x}{A \linkS4class{DiversityIndex} object to be plotted.}
}
\description{
Diversity Plot
}
\examples{
library(magrittr)

\donttest{
## Assemblage diversity size comparison
## Warning: this may take a few seconds!
altamira \%>\%
  as_count() \%>\%
  index_evenness(method = "shannon") \%>\%
  simulate() \%>\%
  plot()

altamira \%>\%
  as_count() \%>\%
  index_richness(method = "none") \%>\%
  simulate() \%>\%
  plot()
}
}
\seealso{
\link{index_heterogeneity}, \link{index_evenness},
 \link{index_richness}

Other plot: 
\code{\link{plot_bar}},
\code{\link{plot_date}()},
\code{\link{plot_line}},
\code{\link{plot_matrix}},
\code{\link{plot_spot}()}
}
\author{
N. Frerebeau
}
\concept{plot}
