% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot_ford.R
\docType{methods}
\name{plot_ford}
\alias{plot_ford}
\alias{plot_ford-method}
\alias{Ford}
\alias{plot_ford,matrix-method}
\alias{plot_ford,data.frame-method}
\title{Ford Diagram}
\usage{
plot_ford(object, ...)

\S4method{plot_ford}{matrix}(
  object,
  weights = FALSE,
  EPPM = FALSE,
  col = "darkgrey",
  axes = TRUE,
  ...
)

\S4method{plot_ford}{data.frame}(object, weights = FALSE, EPPM = FALSE)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} \code{numeric} \code{\link{matrix}} or
\code{\link{data.frame}} of count data (absolute frequencies giving the number of
individuals for each class).}

\item{...}{Currently not used.}

\item{weights}{A \code{\link{logical}} scalar: should row weights (i.e. the number of
observations divided by the total number of observations) be displayed?}

\item{EPPM}{A \code{\link{logical}} scalar: should the EPPM be drawn?
See \code{seriograph()}.}

\item{col}{A vector of colors for the bars.}

\item{axes}{A \code{\link{logical}} scalar: should axes be drawn on the plot?}
}
\value{
\code{plot_ford()} is called it for its side-effects: it results in a graphic
being displayed (invisibly returns \code{object}).
}
\description{
Plots a Ford (battleship curve) diagram.
}
\examples{
## Data from Lipo et al. 2015
data("mississippi", package = "folio")

## Plot a Ford diagram
plot_ford(mississippi)
}
\references{
Ford, J. A. (1962). \emph{A quantitative method for deriving cultural
chronology}. Washington, DC: Pan American Union. Technical manual 1.
}
\seealso{
Other plot methods: 
\code{\link{matrigraph}()},
\code{\link{plot_bertin}()},
\code{\link{plot_diceleraas}()},
\code{\link{plot_diversity}},
\code{\link{plot_heatmap}()},
\code{\link{plot_rank}()},
\code{\link{plot_rarefaction}},
\code{\link{plot_spot}()},
\code{\link{seriograph}()}
}
\author{
N. Frerebeau
}
\concept{plot methods}
