% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot_heatmap.R
\docType{methods}
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap-method}
\alias{plot_heatmap,matrix-method}
\alias{plot_heatmap,data.frame-method}
\alias{plot_heatmap,dist-method}
\title{Heatmap}
\usage{
plot_heatmap(object, ...)

\S4method{plot_heatmap}{matrix}(
  object,
  col = (khroma::colour("YlOrBr"))(12),
  diag = TRUE,
  upper = TRUE,
  lower = TRUE,
  freq = FALSE,
  axes = TRUE,
  legend = TRUE,
  ...
)

\S4method{plot_heatmap}{data.frame}(
  object,
  col = (khroma::colour("YlOrBr"))(12),
  diag = TRUE,
  upper = TRUE,
  lower = TRUE,
  freq = FALSE,
  axes = TRUE,
  legend = TRUE,
  ...
)

\S4method{plot_heatmap}{dist}(
  object,
  col = (khroma::colour("YlOrBr"))(12),
  diag = FALSE,
  upper = FALSE,
  lower = !upper,
  axes = TRUE,
  legend = TRUE,
  ...
)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} \code{numeric} \code{\link{matrix}} or
\code{\link{data.frame}} of count data (absolute frequencies giving the number of
individuals for each class).}

\item{...}{Currently not used.}

\item{col}{A vector of colors.}

\item{diag}{A \code{\link{logical}} scalar indicating whether the diagonal of the
matrix should be plotted. Only used if \code{object} is a symmetric matrix.}

\item{upper}{A \code{\link{logical}} scalar indicating whether the upper triangle of
the matrix should be plotted. Only used if \code{object} is a symmetric matrix.}

\item{lower}{A \code{\link{logical}} scalar indicating whether the lower triangle of
the matrix should be plotted. Only used if \code{object} is a symmetric matrix.}

\item{freq}{A \code{\link{logical}} scalar indicating whether relative frequency
should be used instead of counts (absolute frequency).}

\item{axes}{A \code{\link{logical}} scalar: should axes be drawn on the plot?}

\item{legend}{A \code{\link{logical}} scalar: should a legend be displayed?}
}
\value{
\code{plot_heatmap()} is called it for its side-effects: it results in a graphic
being displayed (invisibly returns \code{object}).
}
\description{
Plots a heatmap.
}
\examples{
## Data from Conkey 1980, Kintigh 1989
data("cantabria")

## Plot matrix diagram...
plot_heatmap(cantabria)
plot_heatmap(cantabria, freq = TRUE)

## Presence/absence data
inc <- sample(0:1, size = 100, replace = TRUE)
bin <- matrix(data = as.logical(inc), nrow = 10, ncol = 10)
plot_heatmap(bin, col = c("black", "white"))
}
\seealso{
Other plot methods: 
\code{\link{matrigraph}()},
\code{\link{plot_bertin}()},
\code{\link{plot_diceleraas}()},
\code{\link{plot_diversity}},
\code{\link{plot_ford}()},
\code{\link{plot_rank}()},
\code{\link{plot_rarefaction}},
\code{\link{plot_spot}()},
\code{\link{seriograph}()}
}
\author{
N. Frerebeau
}
\concept{plot methods}
