% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot_heatmap.R
\docType{methods}
\name{plot_matrix}
\alias{plot_matrix}
\alias{matrigraphe}
\alias{autoplot,matrix-method}
\alias{autoplot,dist-method}
\alias{autoplot,OccurrenceMatrix-method}
\alias{plot,matrix,missing-method}
\alias{plot,dist,missing-method}
\alias{plot,OccurrenceMatrix,missing-method}
\title{Heatmap}
\usage{
\S4method{autoplot}{matrix}(object, ..., diag = TRUE, upper = TRUE, lower = TRUE)

\S4method{autoplot}{dist}(object, ..., diag = FALSE, upper = FALSE, lower = !upper)

\S4method{autoplot}{OccurrenceMatrix}(object, ..., diag = FALSE, upper = FALSE, lower = !upper)

\S4method{plot}{matrix,missing}(x, diag = TRUE, upper = TRUE, lower = TRUE, ...)

\S4method{plot}{dist,missing}(x, diag = FALSE, upper = FALSE, lower = !upper, ...)

\S4method{plot}{OccurrenceMatrix,missing}(x, diag = FALSE, upper = FALSE, lower = !upper, ...)
}
\arguments{
\item{object, x}{A An object to be plotted (typically an object of class
\linkS4class{CountMatrix}).}

\item{...}{Currently not used.}

\item{diag}{A \code{\link{logical}} scalar indicating whether the diagonal of the
matrix should be plotted. Only used if \code{object} is a symmetric matrix.}

\item{upper}{A \code{\link{logical}} scalar indicating whether the upper triangle of
the matrix should be plotted. Only used if \code{object} is a symmetric matrix.}

\item{lower}{A \code{\link{logical}} scalar indicating whether the lower triangle of
the matrix should be plotted. Only used if \code{object} is a symmetric matrix.}
}
\value{
\itemize{
\item \code{autoplot()} returns a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
Plots a heatmap.
}
\examples{
## Abundance data (data from Desachy 2004)
data("compiegne", package = "folio")
## Coerce dataset to absolute frequencies
counts <- as_count(compiegne)
## Coerce dataset to relative frequencies
freq <- as_composition(compiegne)

## Plot matrix diagram...
plot(counts)
plot(freq)

## Presence/absence data
inc <- sample(0:1, size = 100, replace = TRUE)
bin <- IncidenceMatrix(data = inc, nrow = 10, ncol = 10)

autoplot(bin) +
  ggplot2::scale_fill_manual(values = c("TRUE" = "black", "FALSE" = "white"))
}
\references{
Desachy, B. (2004). Le sériographe EPPM: un outil informatisé de sériation
graphique pour tableaux de comptages. \emph{Revue archéologique de Picardie},
3(1), 39-56. \doi{10.3406/pica.2004.2396}.
}
\seealso{
\code{\link[=pvi]{pvi()}}

Other plot: 
\code{\link{plot_bar}},
\code{\link{plot_diversity}},
\code{\link{plot_line}},
\code{\link{plot_spot}()}
}
\author{
N. Frerebeau
}
\concept{plot}
