% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-bar.R
\docType{methods}
\name{plotBar-method}
\alias{plotBar-method}
\alias{seriographe}
\alias{Bertin}
\alias{Ford}
\alias{plotBertin}
\alias{plotBertin-method}
\alias{plotFord}
\alias{plotFord-method}
\alias{plotBertin,CountMatrix-method}
\alias{plotFord,CountMatrix-method}
\title{Bar Plot}
\usage{
plotBertin(object, ...)

plotFord(object, ...)

\S4method{plotBertin}{CountMatrix}(object, threshold = NULL,
  scale = NULL)

\S4method{plotFord}{CountMatrix}(object, level = FALSE, EPPM = FALSE)
}
\arguments{
\item{object}{An object to be plotted.}

\item{...}{Currently not used.}

\item{threshold}{A \code{\link{function}} that takes a numeric vector as
argument and returns a numeric threshold value (see below).
If \code{NULL} (the default), no threshold is computed.}

\item{scale}{A \code{\link{function}} used to scale each variable,
that takes a numeric vector as argument and returns a numeric vector.
If \code{NULL} (the default), no scaling is performed.}

\item{level}{A length-one \code{\link{numeric}} vector giving the
confidence level to be drawn. If \code{NULL} (the default), no confidence
interval is plotted.}

\item{EPPM}{A \code{\link{logical}} scalar: should the EPPM be drawn (see
details)?}
}
\description{
Plots a Bertin or a Ford (battleship curve) diagram.
}
\section{Bertin Matrix}{

 As de Falguerolles \emph{et al.} (1997) points out:
 "In abstract terms, a Bertin matrix is a matrix
 of  displays. [...] To fix ideas, think of a data matrix, variable bycase,
 with real valued variables. For each variable, draw a bar chart of variable
 value by case. High-light all bars representing a value above some sample
 threshold for that variable."
}

\section{Ford Diagram}{

 If \code{EPPM} is \code{TRUE} and if a relative abundance is greater than
 the mean percentage of the type, the exceeding part is highlighted.
 This positive difference from the column mean percentage (in french "écart
 positif au pourcentage moyen", EPPM) represents a deviation from the
 situation of statistical independence. As independence can be interpreted as
 the absence of relationships between types and the chronological order of
 the assemblages, \code{EPPM} is a usefull graphical tool to explore
 significance of relationship between rows and columns related to
 \code{\link[=seriate]{seriation}} (Desachy 2004).
}

\examples{
## Load the mississippi dataset
data("mississippi")
## Coerce the mississippi dataset to a count matrix
mississippi_count <- as(mississippi, "CountMatrix")

## Plot a Bertin diagram...
## ...without threshold
plotBertin(mississippi_count, threshold = NULL)
## ...with the variable mean as threshold
plotBertin(mississippi_count, threshold = mean)

## ...with variables scaled to 0-1
scale_01 <- function(x){
  (x - min(x)) / (max(x) - min(x))
}
plotBertin(mississippi_count, threshold = mean, scale = scale_01)

## Load the mississippi dataset
data("compiegne")
## Coerce the compiegne dataset to a count matrix
compiegne_count <- as(compiegne, "CountMatrix")

## Plot a Ford diagram...
## ...without threshold
plotFord(compiegne_count)
## ...with EPPM
plotFord(compiegne_count, EPPM = TRUE)
}
\references{
Bertin, J. (1977). \emph{La graphique et le traitement graphique de
 l'information}. Paris: Flammarion. Nouvelle Bibliothèque Scientifique.

 de Falguerolles, A., Friedrich, F. & Sawitzki, G. (1997). A Tribute to J.
 Bertin's Graphical Data Analysis. In W. Badilla & F. Faulbaum (eds.),
 \emph{SoftStat '97: Advances in Statistical Software 6}. Stuttgart: Lucius
 & Lucius, p. 11-20.

 Desachy, B. (2004). Le sériographe EPPM: un outil informatisé de sériation
 graphique pour tableaux de comptages. \emph{Revue archéologique de
 Picardie}, 3(1), 39-56.
 DOI: \href{https://doi.org/10.3406/pica.2004.2396}{10.3406/pica.2004.2396}.

 Ford, J. A. (1962). \emph{A quantitative method for deriving cultural
 chronology}. Washington, DC: Pan American Union. Technical manual 1.
}
\seealso{
Other plot: \code{\link{plotDate}}, \code{\link{plotLine}},
  \code{\link{plotMatrix}}, \code{\link{plotSpot}}
}
\author{
N. Frerebeau
}
\concept{plot}
