% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plotDate}
\alias{plotDate}
\alias{plotDate-method}
\alias{plotDate,DateModel,CountMatrix-method}
\alias{plotDate,DateModel,missing-method}
\title{Date plot}
\usage{
plotDate(dates, counts, ...)

\S4method{plotDate}{DateModel,CountMatrix}(dates, counts,
  type = c("event", "accumulation"), select = 1, n = 500)

\S4method{plotDate}{DateModel,missing}(dates, select = 1, sort = "dsc")
}
\arguments{
\item{dates}{An object of class \linkS4class{DateModel} to be plotted.}

\item{counts}{An object of class \linkS4class{CountMatrix} to be plotted.}

\item{...}{Further arguments passed to other methods.}

\item{type}{A \code{\link{character}} string or vector of character strings
indicating the modelled dates to be plotted. It must be one or both
(default) of \code{event} and \code{accumulation}. Any unambiguous substring
can be given.}

\item{select}{A \code{\link{numeric}} or \code{\link{character}} vector
giving the selection of the elements that are drawn.}

\item{n}{A length-one non-negative \code{\link{numeric}} vector giving the
desired length of the vector of quantiles for density computation.}

\item{sort}{A \code{\link{character}} string indicating whether the dates
should be sorted. It can be one of \code{asc} or \code{dsc} (default).
Any unambiguous substring can be given. If \code{NULL} no sorting is
performed.}
}
\description{
Date plot
}
\references{
Bellanger, L. & Husi, P. (2012). Statistical Tool for Dating and
 Interpreting Archaeological Contexts Using Pottery. \emph{Journal of
 Archaeological Science}, 39(4), 777-790.
 DOI: \href{https://doi.org/10.1016/j.jas.2011.06.031}{10.1016/j.jas.2011.06.031}.
}
\seealso{
Other plot: \code{\link{plotBar}},
  \code{\link{plotMatrix}}, \code{\link{plotRank}},
  \code{\link{plotSpot}}
}
\author{
N. Frerebeau
}
\concept{plot}
