% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plotMatrix}
\alias{plotMatrix}
\alias{plotMatrix-method}
\alias{plotMatrix,CountMatrix-method}
\alias{plotMatrix,FrequencyMatrix-method}
\alias{plotMatrix,IncidenceMatrix-method}
\title{Matrix plot}
\usage{
plotMatrix(object, ...)

\S4method{plotMatrix}{CountMatrix}(object, PVI = FALSE, center = FALSE)

\S4method{plotMatrix}{FrequencyMatrix}(object, PVI = FALSE,
  center = FALSE)

\S4method{plotMatrix}{IncidenceMatrix}(object)
}
\arguments{
\item{object}{An object to be plotted.}

\item{...}{Further arguments passed to other methods.}

\item{PVI}{A \code{\link{logical}} scalar: should TODO?}

\item{center}{A \code{\link{logical}} scalar: should TODO?}
}
\description{
XXX
}
\details{
TODO
 coerce to count matrix first
}
\examples{
# Coerce dataset to count matrix
count <- as(compiegne, "CountMatrix")
freq <- as(boves, "FrequencyMatrix")

# Plot matrix diagram...
## ...without threshod (i.e. heatmap)
plotMatrix(count)
plotMatrix(freq)

## ...with PVI as threshold (i.e. Bruno Desachy's matrigraphe)
plotMatrix(count, PVI = TRUE)
plotMatrix(count, PVI = TRUE, center = TRUE)

# Presence/absence data
bin <- IncidenceMatrix(data = sample(0:1, size = 100, replace = TRUE),
                       nrow = 10, ncol = 10)

# Plot matrix diagram
plotMatrix(bin) +
  ggplot2::scale_fill_manual(values = c("TRUE" = "black", "FALSE" = "white"))
}
\references{
Desachy, B. (2004). Le sériographe EPPM: un outil informatisé de sériation
 graphique pour tableaux de comptages. \emph{Revue archéologique de
 Picardie}, 3(1), 39-56.
 DOI: \href{https://doi.org/10.3406/pica.2004.2396}{10.3406/pica.2004.2396}.
}
\seealso{
Other plot: \code{\link{plotBar}}, \code{\link{plotRank}},
  \code{\link{plotSpot}}
}
\author{
N. Frerebeau
}
\concept{plot}
