% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-spot.R
\docType{methods}
\name{plotSpot}
\alias{plotSpot}
\alias{plotSpot-method}
\alias{plotSpot,CountMatrix-method}
\alias{plotSpot,FrequencyMatrix-method}
\alias{plotSpot,SimilarityMatrix-method}
\alias{plotSpot,OccurrenceMatrix-method}
\title{Spot Plot}
\usage{
plotSpot(object, ...)

\S4method{plotSpot}{CountMatrix}(object, threshold = NULL)

\S4method{plotSpot}{FrequencyMatrix}(object, threshold = NULL)

\S4method{plotSpot}{SimilarityMatrix}(object)

\S4method{plotSpot}{OccurrenceMatrix}(object)
}
\arguments{
\item{object}{An object to be plotted.}

\item{...}{Currently not used.}

\item{threshold}{A \code{\link{function}} that takes a numeric vector as
argument and returns a numeric threshold value.
If \code{NULL} (the default), no threshold is computed.}
}
\value{
TODO
}
\description{
Plots a spot matrix.
}
\details{
TODO
 The spot matrix can be considered as a variant of the
 \link[=plotBertin]{Bertin diagram} where the data are first transformed to
 relative frequencies.
}
\note{
Adapted from Dan Gopstein's original
 \href{https://dgopstein.github.io/articles/spot-matrix/}{idea}.
 Credit should be given to him.
}
\examples{
# Plot spot diagram...

## ...of count data...
count <- as(mississippi, "CountMatrix")
### ...without threshod
plotSpot(count)
### ...with the  column means as threshold
plotSpot(count, threshold = mean)
### ...with the column medians as threshold
plotSpot(count, threshold = median)

## ...of a similarity matrix
sim <- similarity(count, "brainerd")
plotSpot(sim)

## ...of a co-occurrence matrix
occ <- as(mississippi, "OccurrenceMatrix")
plotSpot(occ)
}
\seealso{
Other plot: \code{\link{plotBar-method}},
  \code{\link{plotDate}}, \code{\link{plotLine}},
  \code{\link{plotMatrix}}
}
\author{
N. Frerebeau
}
\concept{plot}
