% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-matrix}
\alias{predicate-matrix}
\alias{isSquare}
\alias{isSymmetric}
\title{Matrix predicates}
\usage{
isSquare(x)

isSymmetric(x)
}
\arguments{
\item{x}{A \code{\link{matrix}} to be tested.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
\code{isSquare} checks if a matrix is square.
}
\details{
\code{isSymmetric} checks if a matrix is symmetric.
}
\keyword{internal}
