% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-numeric}
\alias{predicate-numeric}
\alias{isBinary}
\alias{isEqual}
\alias{isIncreasing}
\alias{isDecreasing}
\alias{isOverlapping}
\alias{isPositive}
\alias{isWholeNumber}
\title{Numeric predicates}
\usage{
isBinary(x)

isEqual(x, tolerance = .Machine$double.eps^0.5, na.rm = TRUE)

isIncreasing(x, na.rm = TRUE)

isDecreasing(x, na.rm = TRUE)

isOverlapping(x, y)

isPositive(x, strict = FALSE, na.rm = TRUE)

isWholeNumber(x, tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} object to be tested.}

\item{tolerance}{A \code{\link{numeric}} scalar giving the
tolerance to check within.}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values
(including \code{NaN}) be omitted?}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
Check numeric objects:
}
\details{
\code{isBinary} checks if an object contains only 0s and 1s.

\code{isEqual} checks for equality among all elements of a vector.

\code{isIncreasing} and \code{isDecreasing} check if a sequence of numbers is
monotonically increasing or decreasing.

\code{isOverlapping} checks if two data ranges overlap at all.

\code{isPositive} checks if an object only contains positive values.

\code{isWholeNumber} checks if an object only contains integer (whole)
numbers.
}
\keyword{internal}
