% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-scalar}
\alias{predicate-scalar}
\alias{isScalarList}
\alias{isScalarAtomic}
\alias{isScalarVector}
\alias{isScalarNumeric}
\alias{isScalarInteger}
\alias{isScalarDouble}
\alias{isScalarCharacter}
\alias{isScalarLogical}
\title{Scalar type predicates}
\usage{
isScalarList(x)

isScalarAtomic(x)

isScalarVector(x)

isScalarNumeric(x)

isScalarInteger(x)

isScalarDouble(x)

isScalarCharacter(x)

isScalarLogical(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
Scalar type predicates
}
\keyword{internal}
