% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-type}
\alias{predicate-type}
\alias{isList}
\alias{isAtomic}
\alias{isVector}
\alias{isNumeric}
\alias{isInteger}
\alias{isDouble}
\alias{isCharacter}
\alias{isLogical}
\title{Type predicates}
\usage{
isList(x)

isAtomic(x)

isVector(x)

isNumeric(x)

isInteger(x)

isDouble(x)

isCharacter(x)

isLogical(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
Type predicates
}
\keyword{internal}
