% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-utils}
\alias{predicate-utils}
\alias{isEmpty}
\alias{isSubset}
\title{Utility predicates}
\usage{
isEmpty(x)

isSubset(subset, set)
}
\arguments{
\item{x, y, subset, set}{An object to be tested.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
\code{isEmpty} checks if a vector or list is empty.
\code{isSubset} checks if a vector is a strict subset of another one.
}
\keyword{internal}
