% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pueblo}
\alias{pueblo}
\title{Pueblo IV Period Ceramics}
\format{
A \code{\link{data.frame}} with 9 rows and 5 variables (compositional groups).
}
\source{
Huntley, D. L. (2004). \emph{Interaction, Boundaries, and Identities:
A Multiscalar Approach to the Organizational Scale of Pueblo IV Zuni
Society}. Ph.D. Dissertation, Arizona State University.

Huntley, D. L. (2022). \emph{Ancestral Zuni Glaze-Decorated Pottery: Viewing
Pueblo IV Regional Organization through Ceramic Production and Exchange}.
Anthropological Papers of the University of Arizona 72. Tucson: University
of Arizona Press. \doi{10.2307/j.ctv2ngx5n8}.
}
\usage{
pueblo
}
\description{
A dataset of ceramic counts from the Zuni region.
}
\seealso{
Other datasets: 
\code{\link{aves}},
\code{\link{cantabria}},
\code{\link{woodland}}
}
\concept{datasets}
\keyword{datasets}
