% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RarefactionIndex-class}
\alias{RarefactionIndex-class}
\alias{.RarefactionIndex}
\title{Rarefaction Index}
\description{
An S4 class to represent an rarefaction measure.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A \code{\link{numeric}} matrix of count data (expected taxa per fixed
number of individuals).}

\item{\code{labels}}{A \code{\link{character}} vector giving the sample names.}

\item{\code{size}}{An \code{\link{integer}} vector giving the sample sizes.}

\item{\code{method}}{A \code{\link{character}} string specifying the method used.}
}}

\note{
This class inherits from base \code{\link{matrix}}.
}
\seealso{
\code{\link[=rarefaction]{rarefaction()}}

Other classes: 
\code{\link{CompositionIndex-class}},
\code{\link{DiversityIndex-class}},
\code{\link{EvennessIndex-class}},
\code{\link{HeterogeneityIndex-class}},
\code{\link{RichnessIndex-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
