% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/refine.R
\docType{methods}
\name{refine}
\alias{refine}
\alias{refine-method}
\alias{refine,CountMatrix-method}
\alias{refine,DateModel-method}
\title{Refine models}
\usage{
refine(object, ...)

\S4method{refine}{CountMatrix}(object, cutoff, n = 1000, axes = c(1,
  2), ...)

\S4method{refine}{DateModel}(object, n = 1000, jackknife = TRUE,
  bootstrap = TRUE, ...)
}
\arguments{
\item{object}{An object to refine.}

\item{...}{Further arguments passed to other methods.}

\item{cutoff}{A function that takes a numeric vector as argument and returns
a single numeric value (see details).}

\item{n}{A non-negative \code{\link{integer}} giving the number of partial
bootstrap replications (see details).}

\item{axes}{A \code{\link{numeric}} vector giving the subscripts of the CA
axes to use (see details).}

\item{jackknife}{A \code{\link{logical}} scalar: should the model be checked
by jackknife estimation (removing each fabric/type one at a time)?}

\item{bootstrap}{A \code{\link{logical}} scalar: should the model be checked
by bootstrap resampling?}
}
\description{
See references bellow for more details.
}
\section{Methods (by class)}{
\itemize{
\item \code{CountMatrix}: Performs a partial bootstrap correspondance analysis
 seriation refinement.
 \code{refine} allows to identify samples that are subject to sampling error
 or samples that have underlying structural relationships and might be
 influencing the ordering along the CA space.
 This relies on a partial bootstrap approach to CA-based seriation where each
 sample is replicated \code{n} times. The maximum dimension length of
 the convex hull around the sample point cloud allows to remove samples for
 a given \code{cutoff} value.

 According to Peebles and Schachner (2012), "[this] point removal procedure
 [results in] a reduced dataset where the position of individuals within the
 CA are highly stable and which produces an ordering consistend with the
 assumptions of frequency seriation."

 \code{refine} returns a \linkS4class{BootCA} object containing the subscript
 of samples to be kept (i.e. samples with maximum dimension length of the
 convex hull smaller than the cutoff value).

\item \code{DateModel}: TODO.
}}

\references{
Bellanger, L., Husi, P. & Tomassone, R. (2006). Une approche statistique
 pour la datation de contextes archéologiques. \emph{Revue de Statistique
 Appliquée}, 54(2), 65-81.

 Bellanger, L., Husi, P. & Tomassone, R. (2006). Statistical Aspects of
 Pottery Quantification for the Dating of Some Archaeological Contexts.
 \emph{Archaeometry}, 48(1), 169-183.
 DOI: \href{https://doi.org/10.1111/j.1475-4754.2006.00249.x}{10.1111/j.1475-4754.2006.00249.x}.

 Peeples, M. A., & Schachner, G. (2012). Refining correspondence
 analysis-based ceramic seriation of regional data sets. \emph{Journal of
 Archaeological Science}, 39(8), 2818-2827.
 DOI: \href{https://doi.org/10.1016/j.jas.2012.04.040}{10.1016/j.jas.2012.04.040}.
}
\seealso{
\link{dateEvent} \link{seriate}
}
\author{
N. Frerebeau
}
