% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/alpha-diversity.R
\docType{methods}
\name{richness-method}
\alias{richness-method}
\alias{rarefaction}
\alias{rarefaction-method}
\alias{richness}
\alias{richness,CountMatrix-method}
\alias{richness,IncidenceMatrix-method}
\alias{rarefaction,CountMatrix-method}
\title{Richness and rarefaction}
\usage{
rarefaction(object, ...)

richness(object, ...)

\S4method{richness}{CountMatrix}(object, method = c("margalef",
  "menhinick"), simplify = FALSE)

\S4method{richness}{IncidenceMatrix}(object)

\S4method{rarefaction}{CountMatrix}(object, sample)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} \code{\link{numeric}} matrix.}

\item{...}{Further arguments passed to other methods.}

\item{method}{A \code{\link{character}} string specifiying the index to be
computed. This must be one or more of "\code{margalef}" or
"\code{menhinick}" (see details). Any unambiguous substring can be given.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? The default value, \code{FALSE}, returns a list.}

\item{sample}{A length-one \code{\link{numeric}} vector giving the sub-sample
size.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
\code{richness} returns sample richness. \code{rarefaction} returns
 Hurlbert's unbiaised estimate of Sander's rarefaction.
}
\details{
The number of different taxa, provides an instantly comprehensible
 expression of diversity. While the number of taxa within a sample
 is easy to ascertain, as a term, it makes little sense: some taxa
 may not have been seen, or there may not be a fixed number of taxa
 (e.g. in an open system; Peet 1974). As an alternative, \emph{richness}
 (\eqn{S}) can be used for the concept of taxa number (McIntosh 1967).

 It is not always possible to ensure that all sample sizes are equal
 and the number of different taxa increases with sample size and
 sampling effort. Then, \emph{rarefaction} (\eqn{E(S)}) is the number of
 taxa expected if all samples were of a standard size (i.e. taxa per fixed
 number of individuals). Rarefaction assumes that imbalances between taxa
 are due to sampling and not to differences in actual abundances.

 The following richness measures are available:
 \describe{
  \item{margalef}{Margalef richness index:
   \eqn{D_{Mg} = \frac{S - 1}{\ln N}}{D = (S - 1) / ln N}}
  \item{menhinick}{Menhinick richness index:
   \eqn{D_{Mn} = \frac{S}{\sqrt{N}}}{D = S / \sqrt N}}
 }
}
\note{
Ramanujan approximation is used for \eqn{x!} computation if \eqn{x > 170}.
}
\references{
Hurlbert, S. H. (1971). The Nonconcept of Species Diversity: A Critique and
 Alternative Parameters. \emph{Ecology}, 52(4), 577-586.
 DOI: \href{https://doi.org/10.2307/1934145}{10.2307/1934145}.

 Margalef, R. (1958). Information Theory in Ecology. \emph{General Systems},
 3, 36-71.

 McIntosh, R. P. (1967). An Index of Diversity and the Relation of Certain
 Concepts to Diversity. \emph{Ecology}, 48(3), 392-404.
 DOI: \href{https://doi.org/10.2307/1932674}{10.2307/1932674}.

 Menhinick, E. F. (1964). A Comparison of Some Species-Individuals Diversity
 Indices Applied to Samples of Field Insects. \emph{Ecology}, 45(4), 859-861.
 DOI: \href{https://doi.org/10.2307/1934933}{10.2307/1934933}.

 Peet, R. K. (1974). The Measurement of Species Diversity. \emph{Annual
 Review of Ecology and Systematics}, 5(1), 285-307.
 DOI: \href{https://doi.org/10.1146/annurev.es.05.110174.001441}{10.1146/annurev.es.05.110174.001441}.

 Sander, H. L. (1968). Marine Benthic Diversity: A Comparative Study.
 \emph{The American Naturalist}, 102(925), 243-282.
}
\seealso{
\code{\link[=alpha-diversity]{diversity}}
 \code{\link[=alpha-diversity]{evenness}}
 \code{\link[=beta-diversity]{turnover}}
 \code{\link[=beta-diversity]{similarity}}
}
\author{
N. Frerebeau
}
