% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RichnessIndex-class}
\alias{RichnessIndex-class}
\alias{.RichnessIndex}
\title{Richness Index}
\description{
An S4 class to represent a richness measure.
}
\note{
This class inherits from \linkS4class{DiversityIndex}.
}
\seealso{
\code{\link[=richness]{richness()}}

Other classes: 
\code{\link{CompositionIndex-class}},
\code{\link{DiversityIndex-class}},
\code{\link{EvennessIndex-class}},
\code{\link{HeterogeneityIndex-class}},
\code{\link{RarefactionIndex-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
