% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{roll}
\alias{roll}
\title{Indices of a rolling window}
\usage{
roll(x, window = 3, simplify = FALSE)
}
\arguments{
\item{x}{An object.}

\item{window}{A \code{\link{integer}} scalar giving the window size.}
}
\value{
A \code{\link{list}} with the following components:
 \describe{
  \item{i}{A \code{\link{integer}} vector of indices.}
  \item{w}{A \code{\link{integer}} vector of indices giving the window
  number.}
  \item{m}{A \code{\link{integer}} vector of indices giving the indice of
  the window mid-point.}
 }
}
\description{
Indices of a rolling window
}
\keyword{internal}
