% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{seriate}
\alias{seriate}
\alias{seriate-method}
\title{Seriate}
\usage{
seriate(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Further arguments passed to other methods.}
}
\description{
Seriate
}
\seealso{
\code{\link{permute}}
}
\author{
N. Frerebeau
}
