% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriation.R
\name{seriation-probabilistic}
\alias{seriation-probabilistic}
\alias{seriationReciprocal}
\alias{seriationCorrespondance}
\title{Probabilistic seriation methods}
\usage{
seriationReciprocal(x, margin = 1, stop = 100)

seriationCorrespondance(x, margin, axes, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} matrix.}

\item{margin}{A \code{\link{numeric}} vector giving the subscripts which the
rearrangement will be applied over. E.g., for a matrix \code{1} indicates
rows, \code{2} indicates columns, \code{c(1, 2)} indicates rows then columns,
\code{c(2, 1)} indicates columns then rows.}

\item{stop}{A length-one \code{\link{numeric}} vector giving the stopping rule
(i.e. maximum number of iterations) to avoid infinite loop.}

\item{...}{Further arguments to be passed to \code{\link[FactoMineR]{CA}}.}
}
\value{
A list of two \code{\link{numeric}} vectors.
}
\description{
\code{seriationReciprocal} computes reciprocal ranking/averaging.
\code{seriationCorrespondance} computes CA-based seriation.
}
\author{
N. Frerebeau
}
\concept{seriation methods}
\keyword{internal}
