% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/seriation.R
\docType{methods}
\name{seriation}
\alias{seriation}
\alias{seriate}
\alias{seriate-method}
\alias{permute}
\alias{permute-method}
\alias{seriate,CountMatrix-method}
\alias{permute,CountMatrix,PermutationOrder-method}
\title{Matrix seriation}
\usage{
seriate(object, ...)

permute(object, order, ...)

\S4method{seriate}{CountMatrix}(object, method = c("ranking",
  "correspondance"), EPPM = FALSE, margin = c(1, 2), stop = 100, ...)

\S4method{permute}{CountMatrix,PermutationOrder}(object, order)
}
\arguments{
\item{object}{An object.}

\item{...}{Further arguments passed to other methods.}

\item{order}{TODO.}

\item{method}{A \code{\link{character}} string specifiying the method to be
used. This must be one of "\code{ranking}", "\code{correspondance}"
(see details).
Any unambiguous substring can be given.}

\item{EPPM}{A \code{\link{logical}} scalar: should TODO?.}

\item{margin}{A \code{\link{numeric}} vector giving the subscripts which the
rearrangement will be applied over. E.g., for a matrix \code{1} indicates
rows, \code{2} indicates columns, \code{c(1, 2)} indicates rows then columns,
\code{c(2, 1)} indicates columns then rows.}

\item{stop}{A length-one \code{\link{numeric}} vector giving the stopping rule
(i.e. maximum number of iterations) to avoid infinite loop.}
}
\value{
TODO
}
\description{
\code{seriate} returns permutation order for rows and/or columns.

 \code{permute} returns a rearranged matrix of the same class as
 \code{object}.
}
\references{
Desachy, B. (2004). Le sériographe EPPM: un outil informatisé de sériation
 graphique pour tableaux de comptages. \emph{Revue archéologique de
 Picardie}, 3(1), 39-56.
 DOI: \href{https://doi.org/10.3406/pica.2004.2396}{10.3406/pica.2004.2396}.

 Ihm, P. (2005). A Contribution to the History of Seriation in Archaeology.
 In C. Weihs & W. Gaul (Eds.), \emph{Classification: The Ubiquitous Challenge}
 (p. 307-316). Berlin Heidelberg: Springer.
 DOI: \href{https://doi.org/10.1007/3-540-28084-7_34}{10.1007/3-540-28084-7_34}.
}
\seealso{
\link[FactoMineR]{CA}
}
\author{
N. Frerebeau
}
