% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/index-similarity.R
\docType{methods}
\name{similarity}
\alias{similarity}
\alias{similarity-method}
\alias{similarity,CountMatrix-method}
\alias{similarity,IncidenceMatrix-method}
\title{Similarity}
\usage{
similarity(object, ...)

\S4method{similarity}{CountMatrix}(object, method = c("brainerd", "bray",
  "jaccard", "morisita", "sorenson"), ...)

\S4method{similarity}{IncidenceMatrix}(object, method = c("jaccard",
  "sorenson"), ...)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} matrix of count data.}

\item{...}{Further arguments passed to other methods.}

\item{method}{A \code{\link{character}} string specifiying the method to be
used (see details). Any unambiguous substring can be given.}
}
\value{
\code{similarity} returns a \eqn{m \times m}{m x m} symetric matrix.
}
\description{
\code{similarity} returns a similarity matrix.
}
\details{
\eqn{\beta}-diversity can be measured by addressing \emph{similarity}
 between pairs of sites. This provides a scale of similarity from
 \code{0}-\code{1} where \code{1} is perfect similarity and \code{0} is
 no similarity (with the exception of the Brainerd-Robinson index which is
 scaled between \code{0} and \code{200}):
 \describe{
  \item{brainerd}{Brainerd-Robinson quantitative index. This is a city-block
  metric of similarity.}
  \item{bray}{Sorenson quantitative index (Bray and Curtis modified version
  of the Sorenson index).}
  \item{jaccard}{Jaccard qualitative index.}
  \item{morisita}{Morisita-Horn quantitative index.}
  \item{sorenson}{Sorenson qualitative index.}
 }
}
\examples{
# Data from Magurran 1988, p. 166
birds <- CountMatrix(
  data = c(1.4, 4.3, 2.9, 8.6, 4.2, 15.7, 2.0, 50, 1, 11.4, 11.4, 4.3, 13.0,
           14.3, 8.6, 7.1, 10.0, 1.4, 2.9, 5.7, 1.4, 11.4, 2.9, 4.3, 1.4, 2.9,
           0, 0, 0, 2.9, 0, 0, 0, 10, 0, 0, 5.7, 2.5, 5.7, 8.6, 5.7, 2.9, 0, 0,
           2.9, 0, 0, 5.7, 0, 2.9, 0, 2.9) * 10,
  nrow = 2, byrow = TRUE, dimnames = list(c("unmanaged", "managed"), NULL)
)

# Jaccard measure (presence/absence data)
similarity(birds, "jaccard") # 0.46

# Sorenson measure (presence/absence data)
similarity(birds, "sorenson") # 0.63

# Jaccard measure (Bray's formula ; count data)
similarity(birds, "bray") # 0.44

# Morisita-Horn measure (count data)
similarity(birds, "morisita") # 0.81
}
\references{
Brainerd, G. W. (1951). The Place of Chronological Ordering in
 Archaeological Analysis. \emph{American Antiquity}, 16(04), 301-313.
 DOI: \href{https://doi.org/10.2307/276979}{10.2307/276979}.

 Bray, J. R. & Curtis, J. T. (1957). An Ordination of the Upland Forest
 Communities of Southern Wisconsin. \emph{Ecological Monographs}, 27(4),
 325-349. DOI: \href{https://doi.org/10.2307/1942268}{10.2307/1942268}.

 Magurran, A. E. (1988). \emph{Ecological Diversity and its Measurement}.
 Princeton, NJ: Princeton University Press.
 DOI:\href{https://doi.org/10.1007/978-94-015-7358-0}{10.1007/978-94-015-7358-0}.

 Robinson, W. S. (1951). A Method for Chronologically Ordering Archaeological
 Deposits. \emph{American Antiquity}, 16(04), 293-301.
 DOI: \href{https://doi.org/10.2307/276978}{10.2307/276978}.
}
\seealso{
\code{\link{richness}}
 \code{\link{rarefaction}}
 \code{\link{diversity}}
 \code{\link{evenness}}

Other beta-diversity: \code{\link{turnover}}
}
\author{
N. Frerebeau
}
\concept{beta-diversity}
