% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SimilarityMatrix-class}
\alias{SimilarityMatrix-class}
\title{Similarity matrix}
\arguments{
\item{x}{A \code{SimilarityMatrix} object from which to extract element.}
}
\description{
An S4 class to represent a (dis)similarity matrix.
}
\note{
This class extends the \code{base} \link[base]{matrix}.
}
\seealso{
\link[base]{matrix}
}
\author{
N. Frerebeau
}
