% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/index_diversity.R
\docType{methods}
\name{simulate}
\alias{simulate}
\alias{simulate,DiversityIndex-method}
\title{Measure Diversity by Comparing to Simulated Assemblages}
\usage{
\S4method{simulate}{DiversityIndex}(
  object,
  quantiles = TRUE,
  level = 0.8,
  step = 1,
  n = 1000,
  progress = getOption("tabula.progress")
)
}
\arguments{
\item{object}{A \linkS4class{DiversityIndex} object.}

\item{quantiles}{A \code{\link{logical}} scalar: should sample quantiles be used as
confidence interval? If \code{TRUE} (the default), sample quantiles are used as
described in Kintigh (1984), else quantiles of the t-distribution are
used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{step}{An \code{\link{integer}} giving the increment of the sample size.}

\item{n}{A non-negative \code{\link{integer}} giving the number of bootstrap
replications.}

\item{progress}{A \code{\link{logical}} scalar: should a progress bar be displayed?}
}
\value{
Returns a \linkS4class{DiversityIndex} object.
}
\description{
Measure Diversity by Comparing to Simulated Assemblages
}
\examples{
\donttest{
## Coerce data to a count matrix
data("chevelon", package = "folio")

## Assemblage diversity size comparison
## Warning: this may take a few seconds!
h <- heterogeneity(chevelon, method = "shannon")
h_sim <- simulate(h)
plot(h_sim)

r <- richness(chevelon, method = "count")
r_sim <- simulate(r)
plot(r_sim)
}
}
\references{
Kintigh, K. W. (1984). Measuring Archaeological Diversity by Comparison
with Simulated Assemblages. \emph{American Antiquity}, 49(1), 44-54.
\doi{10.2307/280511}.
}
\seealso{
\code{\link[=plot_diversity]{plot_diversity()}}, \code{\link[=resample]{resample()}}

Other diversity measures: 
\code{\link{heterogeneity}()},
\code{\link{occurrence}()},
\code{\link{plot_diversity}},
\code{\link{rarefaction}()},
\code{\link{richness}()},
\code{\link{similarity}()},
\code{\link{turnover}()}
}
\author{
N. Frerebeau
}
\concept{diversity measures}
