% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabula.R
\docType{package}
\name{tabula}
\alias{tabula}
\alias{tabula-package}
\title{\code{tabula} package}
\description{
Analysis, Seriation and Visualization of Archaeological Count Data
}
\details{
See the README on
\href{https://github.com/nfrerebeau/tabula#readme}{GitHub}
}
