% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/test.R
\docType{methods}
\name{test}
\alias{test}
\alias{test_fit}
\alias{test_fit-method}
\alias{test_fit,CountMatrix-method}
\title{Test}
\usage{
test_fit(object, ...)

\S4method{test_fit}{CountMatrix}(object, method = c("FIT"),
  simplify = FALSE, ...)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} matrix of count data.}

\item{...}{Further arguments to be passed to internal methods.}

\item{method}{A \code{\link{character}} string specifiying the method to be
used (see details). Any unambiguous substring can be given.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix? The default value, \code{FALSE}, returns a list.}
}
\value{
TODO
}
\description{
Test
}
\details{
The following methods are available:
 \describe{
  \item{FIT}{The Frequency Increment Test (Feder et al. 2014). This
  test rejects neutrality if the distribution of normalized variant frequency
  increments exhibits a mean that deviates significantly from zero.}
 }
}
\examples{
## Load the merzbach dataset
data("merzbach")

## Coerce the merzbach dataset to a count matrix
## Keep only decoration types that have a maximum frequency of at least 50
keep <- apply(X = merzbach, MARGIN = 2, FUN = function(x) max(x) >= 50)
count_merzbach <- as(merzbach[, keep], "CountMatrix")

## The data are grouped by phase
## We use the row names as time coordinates (roman numerals)
set_dates(count_merzbach) <- rownames(merzbach)

fit <- test_fit(count_merzbach, simplify = TRUE)
}
\references{
Feder, A. F., Kryazhimskiy, S. & Plotkin, J. B. (2014). Identifying
 Signatures of Selection in Genetic Time Series. \emph{Genetics}, 196(2),
 509-522.
 DOI: \href{https://doi.org/10.1534/genetics.113.158220}{10.1534/genetics.113.158220}.
}
\author{
N. Frerebeau
}
