% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/test_fit.R
\docType{methods}
\name{test_fit}
\alias{test_fit}
\alias{test_fit-method}
\alias{test_fit,CountMatrix-method}
\title{Frequency Increment Test}
\usage{
test_fit(object, ...)

\S4method{test_fit}{CountMatrix}(object, simplify = FALSE, ...)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} matrix of count data.}

\item{...}{Currently not used.}

\item{simplify}{A \code{\link{logical}} scalar: should the result be
simplified to a matrix?}
}
\value{
If \code{simplify} is \code{FALSE}, returns a list (default), else returns
 a matrix.
}
\description{
Frequency Increment Test
}
\details{
The Frequency Increment Test (FIT) rejects neutrality if the distribution
 of normalized variant frequency increments exhibits a mean that deviates
 significantly from zero.
}
\examples{
## Shannon diversity test
merzbach_count <- as_count(merzbach)
div <- test_diversity(merzbach_count)

## Frequency Increment Test
## Coerce the merzbach dataset to a count matrix
## Keep only decoration types that have a maximum frequency of at least 50
keep <- apply(X = merzbach, MARGIN = 2, FUN = function(x) max(x) >= 50)
merzbach_count <- as_count(merzbach[, keep])

## The data are grouped by phase
## We use the row names as time coordinates (roman numerals)
set_dates(merzbach_count) <- rownames(merzbach)

fit <- test_fit(merzbach_count, simplify = TRUE)
}
\references{
Feder, A. F., Kryazhimskiy, S. & Plotkin, J. B. (2014). Identifying
 Signatures of Selection in Genetic Time Series. \emph{Genetics}, 196(2),
 509-522.
 DOI: \href{https://doi.org/10.1534/genetics.113.158220}{10.1534/genetics.113.158220}.
}
\seealso{
Other statistics: 
\code{\link{refine}},
\code{\link{test_diversity}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
