% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woodland}
\alias{woodland}
\title{Trees Incidences}
\format{
A \code{\link{data.frame}} with 6 rows (quadarts) and 6 variables (tree species).
}
\source{
Magurran, A. E. (1988). \emph{Ecological Diversity and its Measurement}.
Princeton, NJ: Princeton University Press.
\doi{10.1007/978-94-015-7358-0}.
}
\usage{
woodland
}
\description{
A dataset of presence or absence of trees in six (10 x 10 m) quadarts along
a transect through a deciduous woodland.
}
\seealso{
Other datasets: 
\code{\link{aves}},
\code{\link{cantabria}},
\code{\link{pueblo}}
}
\concept{datasets}
\keyword{datasets}
