% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zuni}
\alias{zuni}
\title{Zuni ceramics}
\format{A numeric matrix with 420 rows and 18 variables:
 \describe{
   \item{LINO}{}
   \item{KIAT}{}
   \item{RED}{}
   \item{GALL}{}
   \item{ESC}{}
   \item{PUBW}{}
   \item{RES}{}
   \item{TULA}{}
   \item{PINE}{}
   \item{PUBR}{}
   \item{WING}{}
   \item{WIPO}{}
   \item{SJ}{}
   \item{LSJ}{}
   \item{SPR}{}
   \item{PINER}{}
   \item{HESH}{}
   \item{KWAK}{}
 }}
\source{
Peeples, M. A., & Schachner, G. (2012). Refining correspondence
 analysis-based ceramic seriation of regional data sets. \emph{Journal of
 Archaeological Science}, 39(8), 2818-2827.
 DOI: \href{https://doi.org/10.1016/j.jas.2012.04.040}{10.1016/j.jas.2012.04.040}.
}
\usage{
zuni
}
\description{
A dataset containing ceramic counts from the Zuni region of the American
Southwest.
}
\keyword{datasets}
