% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/extract.R
\docType{methods}
\name{access}
\alias{access}
\alias{get}
\alias{set}
\alias{getID}
\alias{getTotals}
\alias{getDates}
\alias{setDates<-}
\alias{getCoordinates}
\alias{setCoordinates<-}
\alias{getEPSG}
\alias{setEPSG<-}
\alias{getCoordinates,AbundanceMatrix-method}
\alias{getDates,AbundanceMatrix-method}
\alias{getEPSG,AbundanceMatrix-method}
\alias{getID,ANY-method}
\alias{getTotals,FrequencyMatrix-method}
\alias{setDates<-,AbundanceMatrix-method}
\alias{setDates,AbundanceMatrix-method}
\alias{setCoordinates<-,AbundanceMatrix-method}
\alias{setCoordinates,AbundanceMatrix-method}
\alias{setEPSG<-,AbundanceMatrix-method}
\alias{setEPSG,AbundanceMatrix-method}
\title{Extract or Replace Parts of an Object}
\usage{
getID(object)

getTotals(object)

getDates(object)

setDates(object) <- value

getCoordinates(object)

setCoordinates(object) <- value

getEPSG(object)

setEPSG(object) <- value

\S4method{getCoordinates}{AbundanceMatrix}(object)

\S4method{getDates}{AbundanceMatrix}(object)

\S4method{getEPSG}{AbundanceMatrix}(object)

\S4method{getID}{ANY}(object)

\S4method{getTotals}{FrequencyMatrix}(object)

\S4method{setDates}{AbundanceMatrix}(object) <- value

\S4method{setCoordinates}{AbundanceMatrix}(object) <- value

\S4method{setEPSG}{AbundanceMatrix}(object) <- value
}
\arguments{
\item{object}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{object} (see
below).}
}
\description{
Getters and setters to extract or replace parts of an object.
}
\section{Set dates}{

 An attempt is made to interpret the argument \code{value} in a suitable way.
 If \code{value} is a character vector, it is assumed to contain Roman
 numerals. \emph{Note} that errors are assumed to be given at \code{1} sigma.
 If \code{value} is a:
 \describe{
  \item{\code{numeric} or \code{integer} \code{vector}}{these values are
  assumed to represent dates without known errors.}
  \item{\code{list}}{containing components "\code{value}" and "\code{error}",
  these are used to define dates and corresponding errors.}
  \item{\code{matrix} or \code{data.frame} with two or more columns}{the
  first is assumed to contain the dates and the second the error values.
  \emph{Note} that if \code{value} has columns named "\code{value}" and
  "\code{error}", these columns will be used.}
 }
}

\section{Set coordinates}{

 An attempt is made to interpret the argument \code{value} in a way suitable
 for geographic coordinates. If \code{value} is a:
 \describe{
  \item{\code{list}}{containing components "\code{x}", "\code{y}" and
  "\code{z}", these are used to define coordinates (longitude, latitude and
  elevation, respectively). If "\code{z}" is missing, the vertical
  coordinates will be ignored (and \code{NA} will be generated).}
  \item{\code{matrix} or \code{data.frame} with two or more columns}{the
  first is assumed to contain the \code{x} values, the second the \code{y}
  and the third the \code{z} values. \emph{Note} that if \code{value} has
  columns named "\code{x}", "\code{y}" and "\code{z}", these columns will be
  used. If \code{value} has only two columns or has columns named "\code{x}"
  and "\code{y}" but not "\code{z}", the vertical coordinates will be ignored
  (and \code{NA} will be generated).}
 }
}

\author{
N. Frerebeau
}
