% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/extract.R
\docType{methods}
\name{access}
\alias{access}
\alias{get}
\alias{set}
\alias{get_id}
\alias{get_id-method}
\alias{get_totals}
\alias{get_totals-method}
\alias{set_totals<-}
\alias{set_totals-method}
\alias{get_id,ANY-method}
\alias{get_totals,FrequencyMatrix-method}
\alias{set_totals<-,FrequencyMatrix-method}
\alias{set_totals,FrequencyMatrix-method}
\title{Get or Set Parts of an Object}
\usage{
get_id(object)

get_totals(object)

set_totals(object) <- value

\S4method{get_id}{ANY}(object)

\S4method{get_totals}{FrequencyMatrix}(object)

\S4method{set_totals}{FrequencyMatrix}(object) <- value
}
\arguments{
\item{object}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{object} (see
below).}
}
\value{
An object of the same sort as \code{object} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\author{
N. Frerebeau
}
