% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/extract.R
\docType{methods}
\name{access}
\alias{access}
\alias{get}
\alias{set}
\alias{get_id}
\alias{get_totals}
\alias{get_dates}
\alias{set_dates<-}
\alias{get_coordinates}
\alias{set_coordinates<-}
\alias{get_epsg}
\alias{set_epsg<-}
\alias{get_coordinates,AbundanceMatrix-method}
\alias{get_dates,AbundanceMatrix-method}
\alias{get_epsg,AbundanceMatrix-method}
\alias{get_id,ANY-method}
\alias{get_totals,FrequencyMatrix-method}
\alias{set_dates<-,AbundanceMatrix-method}
\alias{set_dates,AbundanceMatrix-method}
\alias{set_coordinates<-,AbundanceMatrix-method}
\alias{set_coordinates,AbundanceMatrix-method}
\alias{set_epsg<-,AbundanceMatrix-method}
\alias{set_epsg,AbundanceMatrix-method}
\title{Extract or Replace Parts of an Object}
\usage{
get_id(object)

get_totals(object)

get_dates(object)

set_dates(object) <- value

get_coordinates(object)

set_coordinates(object) <- value

get_epsg(object)

set_epsg(object) <- value

\S4method{get_coordinates}{AbundanceMatrix}(object)

\S4method{get_dates}{AbundanceMatrix}(object)

\S4method{get_epsg}{AbundanceMatrix}(object)

\S4method{get_id}{ANY}(object)

\S4method{get_totals}{FrequencyMatrix}(object)

\S4method{set_dates}{AbundanceMatrix}(object) <- value

\S4method{set_coordinates}{AbundanceMatrix}(object) <- value

\S4method{set_epsg}{AbundanceMatrix}(object) <- value
}
\arguments{
\item{object}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{object} (see
below).}
}
\value{
TODO
}
\description{
Getters and setters to extract or replace parts of an object.
}
\section{Set dates}{

 An attempt is made to interpret the argument \code{value} in a suitable way.
 \emph{Note} that errors are assumed to be given at \code{1} sigma.

 If \code{value} is a:
 \describe{
  \item{\code{character} vector}{it is assumed to contain Roman numerals.}
  \item{\code{numeric} or \code{integer} \code{vector}}{these values are
  assumed to represent dates without known errors.}
  \item{\code{list}}{containing components "\code{value}" and "\code{error}",
  these are used to define dates and corresponding errors.}
  \item{\code{matrix} or \code{data.frame} with two or more columns}{the
  first is assumed to contain the dates and the second the error values.
  \emph{Note} that if \code{value} has columns named "\code{value}" and
  "\code{error}", these columns will be used.}
 }
}

\section{Set coordinates}{

 An attempt is made to interpret the argument \code{value} in a way suitable
 for geographic coordinates. If \code{value} is a:
 \describe{
  \item{\code{list}}{containing components "\code{x}", "\code{y}" and
  "\code{z}", these are used to define coordinates (longitude, latitude and
  elevation, respectively). If "\code{z}" is missing, the vertical
  coordinates will be ignored (and \code{NA} will be generated).}
  \item{\code{matrix} or \code{data.frame} with two or more columns}{the
  first is assumed to contain the \code{x} values, the second the \code{y}
  and the third the \code{z} values. \emph{Note} that if \code{value} has
  columns named "\code{x}", "\code{y}" and "\code{z}", these columns will be
  used. If \code{value} has only two columns or has columns named "\code{x}"
  and "\code{y}" but not "\code{z}", the vertical coordinates will be ignored
  (and \code{NA} will be generated).}
 }
}

\author{
N. Frerebeau
}
