% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\docType{methods}
\name{accessors}
\alias{accessors}
\alias{accumulation}
\alias{columns}
\alias{rows}
\alias{totals}
\title{Accessors}
\usage{
accumulation(x)

columns(x)

rows(x)

totals(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Accessors
}
\author{
N. Frerebeau
}
