% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{accessors}
\alias{accessors}
\alias{columns}
\alias{method}
\alias{rows}
\alias{totals}
\title{Accessors}
\usage{
columns(x)

method(x)

rows(x)

totals(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Accessors
}
\author{
N. Frerebeau
}
