% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{accessors}
\alias{accessors}
\alias{totals}
\title{Accessors}
\usage{
totals(x)
}
\arguments{
\item{x}{An object from which to extract element(s).}
}
\description{
Accessors
}
\author{
N. Frerebeau
}
