% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BootDate-class}
\alias{BootDate-class}
\title{Date model checking}
\description{
An S4 class to store the results of resampling methods for date model
 checking.
}
\section{Slots}{

\describe{
\item{\code{jackknife}}{A six columns \code{\link{data.frame}} giving the results of
the resamping procedure (jackknifing) for each sample (in rows) with the
following columns:
\describe{
 \item{id}{An identifier.}
 \item{estimation}{The jackknife date estimate.}
 \item{earliest}{The lower boundary of the associated prediction interval.}
 \item{latest}{The upper boundary of the associated prediction interval.}
 \item{error}{The standard error of predicted means.}
 \item{bias}{The jackknife estimate of bias.}
}}

\item{\code{bootstrap}}{A six columns \code{\link{data.frame}} giving the boostrap
distribution statistics for each replicated sample (in rows) with the
following columns:
\describe{
 \item{id}{An identifier.}
 \item{min}{Minimum value.}
 \item{Q05}{Sample quantile to 0.05 probability.}
 \item{mean}{Mean value.}
 \item{Q95}{Sample quantile to 0.95 probability.}
 \item{max}{Maximum value.}
}}
}}

\author{
N. Frerebeau
}
