% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot_ca.R
\docType{methods}
\name{ca}
\alias{ca}
\alias{plot_ca}
\alias{plot_ca-method}
\alias{plot_ca,CA-method}
\alias{plot_ca,BootCA-method}
\title{Correspondence Analysis}
\usage{
plot_ca(object, ...)

\S4method{plot_ca}{CA}(object, axes = c(1, 2), map = c("rows", "columns"), ...)

\S4method{plot_ca}{BootCA}(object, axes = c(1, 2), map = c("rows", "columns"), keep = NULL, ...)
}
\arguments{
\item{object}{A \linkS4class{CA} object.}

\item{...}{Currently not used.}

\item{axes}{A length-two \code{\link{numeric}} vector specifying the
components to plot.}

\item{map}{A \code{\link{character}} vector specifying the graph to plot.
It must be one or both of "\code{rows}" or "\code{columns}".
Any unambiguous substring can be given.}

\item{keep}{A \code{\link{character}} vector specifying the subset to plot.
It must be \code{NULL} or one or both of "\code{rows}" or "\code{columns}".
(see \code{\link{refine_ca}}).}
}
\description{
Plots results of a correspondence analysis (or reciprocal averaging).
}
\examples{
## Coerce dataset to a count matrix
counts <- as_count(compiegne)

## Computes correspondence analysis
ca <- arkhe::ca(counts)
plot_ca(ca, map = c("rows", "columns"))

}
\seealso{
\link{refine_ca}, \link{seriate_correspondence}
}
\author{
N. Frerebeau
}
\concept{mutlivariate}
