% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{as_matrix}
\alias{as_matrix-method}
\alias{as_matrix,matrix-method}
\alias{as_matrix,data.frame-method}
\title{Coerce}
\usage{
as_matrix(from)

\S4method{as_matrix}{matrix}(from)

\S4method{as_matrix}{data.frame}(from)
}
\arguments{
\item{from}{A numeric \code{\link{matrix}} or \code{\link{data.frame}} to be
coerced.}
}
\description{
Coerce
}
\author{
N. Frerebeau
}
