% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dating.R
\docType{methods}
\name{dateEvent}
\alias{dateEvent}
\alias{dateEvent-method}
\alias{dateAccumulation}
\alias{dateAccumulation-method}
\alias{dateEvent,CountMatrix-method}
\alias{dateAccumulation,CountMatrix-method}
\title{Date archaeological assemblages}
\usage{
dateEvent(object, ...)

dateAccumulation(object, ...)

\S4method{dateEvent}{CountMatrix}(object, time, axes = 5, level = 0.95,
  verbose = FALSE, ...)

\S4method{dateAccumulation}{CountMatrix}(object, time, axes = 5,
  level = 0.95, verbose = FALSE, ...)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} matrix of count data.}

\item{...}{Further arguments to be passed to \code{\link[FactoMineR]{CA}}.}

\item{time}{A named \code{\link{numeric}} vector giving the reliable known
dates (in years) of the archaeological assemblages.
Assemblages are matched by names.}

\item{axes}{An \code{\link{integer}} indicating th factorial components of
the correspondance analysis to retain (by default 5).}

\item{level}{A length-one \code{\link{numeric}} vector giving the
confidence level.}

\item{verbose}{A \code{\link{logical}} scalar. Should extra summary
statistics be printed?}
}
\value{
An object of class \linkS4class{DateEvent} or \linkS4class{DateAccumulation}.
}
\description{
Date archaeological assemblages
}
\references{
Bellanger, L. & Husi, P. (2012). Statistical Tool for Dating and
 Interpreting Archaeological Contexts Using Pottery. \emph{Journal of
 Archaeological Science}, 39(4), 777-790.
 DOI: \href{https://doi.org/10.1016/j.jas.2011.06.031}{10.1016/j.jas.2011.06.031}.
}
\author{
N. Frerebeau
}
