% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dating.R
\docType{methods}
\name{dateEvent}
\alias{dateEvent}
\alias{dateEvent-method}
\alias{dateEvent,CountMatrix-method}
\title{Date archaeological assemblages}
\usage{
dateEvent(object, ...)

\S4method{dateEvent}{CountMatrix}(object, dates, axes = 5,
  level = 0.95, verbose = FALSE, ...)
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} matrix of count data.}

\item{...}{Further arguments to be passed to \code{\link[FactoMineR]{CA}}.}

\item{dates}{A named \code{\link{numeric}} vector giving the reliable known
dates (in years) of the archaeological assemblages.
Assemblages are matched by names.}

\item{axes}{An \code{\link{integer}} indicating the factorial components of
the correspondance analysis to retain (by default 5).}

\item{level}{A length-one \code{\link{numeric}} vector giving the
confidence level.}

\item{verbose}{A \code{\link{logical}} scalar. Should extra summary
statistics be printed?}
}
\value{
An object of class \linkS4class{DateEvent}.
}
\description{
Date archaeological assemblages
}
\examples{
zuni <- as(zuni, "CountMatrix")

dates <- c(
  LZ0569 = 1097, LZ0279 = 1119, CS16 = 1328, LZ0066 = 1111,
  LZ0852 = 1216, LZ1209 = 1251, CS144 = 1262, LZ0563 = 1206,
  LZ0329 = 1076, LZ0005Q = 859, LZ0322 = 1109, LZ0067 = 863,
  LZ0578 = 1180, LZ0227 = 1104, LZ0610 = 1074
)

(events <- dateEvent(zuni, dates, axes = 10, verbose = TRUE))

head(events[["rows"]])

events[["columns"]]

plotDate(zuni, events, select = 1:5) +
  ggplot2::theme_bw()

plotBar(events, select = 1:20) +
  ggplot2::theme_bw()
}
\references{
Bellanger, L. & Husi, P. (2012). Statistical Tool for Dating and
 Interpreting Archaeological Contexts Using Pottery. \emph{Journal of
 Archaeological Science}, 39(4), 777-790.
 DOI: \href{https://doi.org/10.1016/j.jas.2011.06.031}{10.1016/j.jas.2011.06.031}.
}
\author{
N. Frerebeau
}
