% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{DateModel-class}
\alias{DateModel-class}
\alias{[[,DateModel-method}
\alias{[,DateModel-method}
\title{Date model}
\usage{
\S4method{[[}{DateModel}(x, i)

\S4method{[}{DateModel}(x, i, j)
}
\arguments{
\item{x}{A \code{DateModel} object from which to extract element(s).}

\item{i, j}{Indices specifying elements to extract.
\code{i} is a \code{\link{character}} string matching to the name of a slot.
\code{j} can be \code{\link{missing}} or \code{\link{NULL}},
a \code{\link{numeric}} or \code{\link{character}} vector.
Numeric values are coerced to \code{\link{integer}} as by
\code{\link{as.integer}} (and hence truncated towards zero).
Character vectors will be matched to the names of the object.}
}
\description{
An S4 class to store the event and accumulation times of archaeological
 assemblages.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[[}: Extracts informations from a slot selected by subscript
\code{i}.

\item \code{[}: Allows to select a slot thru \code{j}.
}}

\section{Slots}{

\describe{
\item{\code{counts}}{A \eqn{m \times p}{m x p} \code{\link{numeric}} matrix of count
data used for model fitting.}

\item{\code{dates}}{A two columns \code{\link{data.frame}} giving the known dates
used for model fitting and an identifier to link each row to an assemblage.}

\item{\code{level}}{A length-one \code{\link{numeric}} vector giving the
confidence level.}

\item{\code{model}}{A \code{\link[stats:lm]{multiple linear model}}: the Gaussian
multiple linear regression model fitted for event date estimation and
prediction.}

\item{\code{residual}}{A length-one \code{\link{numeric}} vector giving the residual
standard deviation.}

\item{\code{rows}}{A five columns \code{\link{data.frame}} giving the predicted event
dates for each archaeological assemblage, the corresponding confidence
interval boundaries, standard error of the predicted dates and an
identifier to link each row to an assemblage.}

\item{\code{columns}}{A five columns \code{\link{data.frame}} giving the predicted
event dates for each archaeological type or fabric, the corresponding
confidence interval boundaries, standard error of the predicted dates and an
identifier to link each row to a type or fabric.}

\item{\code{accumulation}}{A two columns \code{\link{data.frame}} giving the point
estimate of accumulation dates of archaeological assemblages and an
identifier to link each row to an assemblage.}
}}

\author{
N. Frerebeau
}
