% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecate.R
\docType{methods}
\name{deprecate}
\alias{deprecate}
\alias{seriate_correspondence}
\alias{seriate_correspondence-method}
\alias{seriate_reciprocal}
\alias{seriate_reciprocal-method}
\alias{seriate_reciprocal,CountMatrix-method}
\alias{seriate_reciprocal,IncidenceMatrix-method}
\title{Deprecated Methods}
\usage{
seriate_correspondence(object, ...)

seriate_reciprocal(object, ...)

\S4method{seriate_reciprocal}{CountMatrix}(object, EPPM = FALSE, margin = c(1, 2), stop = 100)

\S4method{seriate_reciprocal}{IncidenceMatrix}(object, margin = c(1, 2), stop = 100)
}
\arguments{
\item{object}{An object.}

\item{...}{Currently not used.}

\item{EPPM}{A \code{\link{logical}} scalar: should the seriation be computed
on EPPM instead of raw data?}

\item{margin}{A \code{\link{numeric}} vector giving the subscripts which the
rearrangement will be applied over: \code{1} indicates rows, \code{2}
indicates columns, \code{c(1, 2)} indicates rows then columns,
\code{c(2, 1)} indicates columns then rows.}

\item{stop}{An \code{\link{integer}} giving the stopping rule
(i.e. maximum number of iterations) to avoid infinite loop.}
}
\description{
Deprecated Methods
}
\author{
N. Frerebeau
}
