% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-bar.R, R/plot-matrix.R,
%   R/plot-rank.R, R/plot-spot.R
\docType{methods}
\name{tabula-deprecated}
\alias{tabula-deprecated}
\alias{plotBar}
\alias{plotMatrix}
\alias{plotRank}
\alias{plotSpot}
\alias{plotBar,CountMatrix-method}
\alias{plotBar,FrequencyMatrix-method}
\alias{plotMatrix,CountMatrix-method}
\alias{plotMatrix,FrequencyMatrix-method}
\alias{plotMatrix,IncidenceMatrix-method}
\alias{plotRank,CountMatrix-method}
\alias{plotRank,FrequencyMatrix-method}
\alias{plotSpot,CountMatrix-method}
\alias{plotSpot,FrequencyMatrix-method}
\alias{plotSpot,SimilarityMatrix-method}
\alias{plotSpot,OccurrenceMatrix-method}
\title{Deprecated Methods}
\usage{
plotBar(object, ...)

plotMatrix(object, ...)

plotRank(object, ...)

plotSpot(object, ...)

\S4method{plotBar}{CountMatrix}(object, level = FALSE, EPPM = FALSE,
  center = TRUE, horizontal = FALSE)

\S4method{plotBar}{FrequencyMatrix}(object, level = FALSE,
  EPPM = FALSE, center = TRUE, horizontal = FALSE)

\S4method{plotMatrix}{CountMatrix}(object, PVI = FALSE)

\S4method{plotMatrix}{FrequencyMatrix}(object, PVI = FALSE)

\S4method{plotMatrix}{IncidenceMatrix}(object)

\S4method{plotRank}{CountMatrix}(object, PVI = FALSE)

\S4method{plotRank}{FrequencyMatrix}(object, PVI = FALSE)

\S4method{plotSpot}{CountMatrix}(object, threshold = NULL)

\S4method{plotSpot}{FrequencyMatrix}(object, threshold = NULL)

\S4method{plotSpot}{SimilarityMatrix}(object)

\S4method{plotSpot}{OccurrenceMatrix}(object)
}
\arguments{
\item{object}{An object to be plotted.}

\item{...}{Further arguments to be passed to internal methods.}

\item{level}{A length-one \code{\link{numeric}} vector giving the
confidence level to be drawn.}

\item{EPPM}{A \code{\link{logical}} scalar: should the EPPM be drawn (see
details)?}

\item{center}{A \code{\link{logical}} scalar: should the bar plot
be centered? The default, \code{TRUE}, produces a Ford diagram, otherwise it
produces a Bertin diagram.}

\item{horizontal}{A \code{\link{logical}} scalar: should the bar plot
be horizontal? The default, \code{FALSE}, means variables in rows and
cases in columns (i.e. Bertin diagram). Only used if \code{center} is
\code{FALSE}.}

\item{PVI}{A \code{\link{logical}} scalar: should the PVI be drawn instead of
frequencies (see details)?}

\item{log}{A \code{\link{character}} string which contains "\code{x}" if the
x axis is to be logarithmic, "\code{y}" if the y axis is to be logarithmic
and "\code{xy}" or "\code{yx}" if both axes are to be logarithmic (base 10).}

\item{facet}{A \code{\link{logical}} scalar: should a matrix of panels
defined by case/sample be drawn?}
}
\value{
TODO
}
\description{
\code{plotBar} produces a Bertin or a Ford (battleship curve) diagram.
}
\keyword{internal}
