% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-bar.R
\docType{methods}
\name{deprecated}
\alias{deprecated}
\alias{plotBar}
\alias{plotBar,CountMatrix-method}
\alias{plotBar,FrequencyMatrix-method}
\title{Deprecated Methods}
\usage{
plotBar(object, ...)

\S4method{plotBar}{CountMatrix}(object, level = FALSE, EPPM = FALSE,
  center = TRUE, horizontal = FALSE)

\S4method{plotBar}{FrequencyMatrix}(object, level = FALSE,
  EPPM = FALSE, center = TRUE, horizontal = FALSE)
}
\arguments{
\item{object}{An object to be plotted.}

\item{...}{Further arguments passed to other methods.}

\item{level}{A length-one \code{\link{numeric}} vector giving the
confidence level to be drawn.}

\item{EPPM}{A \code{\link{logical}} scalar: should the EPPM be drawn (see
details)?}

\item{center}{A \code{\link{logical}} scalar: should the bar plot
be centered? The default, \code{TRUE}, produces a Ford diagram, otherwise it
produces a Bertin diagram.}

\item{horizontal}{A \code{\link{logical}} scalar: should the bar plot
be horizontal? The default, \code{FALSE}, means variables in rows and
cases in columns (i.e. Bertin diagram). Only used if \code{center} is
\code{FALSE}.}
}
\description{
\code{plotBar} produces a Bertin or a Ford (battleship curve) diagram.
}
\keyword{internal}
