% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DiversityIndex}
\alias{DiversityIndex}
\alias{DiversityIndex-class}
\alias{.DiversityIndex}
\alias{HeterogeneityIndex-class}
\alias{.HeterogeneityIndex}
\alias{EvennessIndex-class}
\alias{.EvennessIndex}
\alias{RichnessIndex-class}
\alias{.RichnessIndex}
\alias{CompositionIndex-class}
\alias{.CompositionIndex}
\title{Diversity Index}
\description{
An S4 class to represent a diversity measure.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{\link{numeric}} matrix of count data.}

\item{\code{values}}{A \code{\link{numeric}} vector giving the diversity index values.}

\item{\code{size}}{A \code{\link{integer}} vector giving the sample sizes.}

\item{\code{simulation}}{A \code{\link{numeric}} matrix giving the diversity
measures for the simulated assemblage.}

\item{\code{method}}{A \code{\link{character}} string specifying the method used.}
}}

\section{Subset}{

 In the code snippets below, \code{x} is a \code{DiversityIndex} object.
 \describe{
  \item{\code{x[[i]]}}{Extracts informations from a slot selected by
  subscript \code{i}. \code{i} is a length-one \code{\link{character}}
  vector. Returns the corresponding slot values.}
 }
}

\seealso{
Other class: 
\code{\link{DateEvent-class}},
\code{\link{DateMCD-class}},
\code{\link{IncrementTest-class}},
\code{\link{PermutationOrder-class}},
\code{\link{RefineCA-class}}
}
\author{
N. Frerebeau
}
\concept{class}
