% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resample.R
\docType{methods}
\name{jackknife}
\alias{jackknife}
\alias{jackknife,DiversityIndex-method}
\title{Jackknife Estimation}
\usage{
\S4method{jackknife}{DiversityIndex}(object, f = NULL)
}
\arguments{
\item{object}{An \R object (typically a \linkS4class{DiversityIndex} object).}

\item{f}{A \code{\link{function}} that takes a single numeric vector (the leave-one-out
values) as argument.}
}
\value{
If \code{f} is \code{NULL} (the default), \code{jackknife()} returns a \code{numeric} \code{matrix}
with the following columns:
\describe{
\item{\code{original}}{The observed value.}
\item{\code{mean}}{The jackknife estimate of mean.}
\item{\code{bias}}{The jackknife estimate of bias.}
\item{\code{error}}{The jackknife estimate of standard error.}
}

If \code{f} is a \code{function}, \code{jackknife()} returns the result of \code{f} applied to
the leave-one-out values.
}
\description{
Jackknife Estimation
}
\examples{
## Data from Conkey 1980, Kintigh 1989
data("cantabria")

## Shannon diversity index
(h <- heterogeneity(cantabria, method = "shannon"))

## Jackknife resampling
jackknife(h)

jackknife(h, f = summary)
}
\seealso{
Other resampling methods: 
\code{\link{bootstrap}()}
}
\author{
N. Frerebeau
}
\concept{resampling methods}
