% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutators}
\alias{mutators}
\alias{get}
\alias{set}
\alias{get_method}
\alias{get_method-method}
\alias{labels,DiversityIndex-method}
\alias{labels,RarefactionIndex-method}
\alias{get_method,DiversityIndex-method}
\title{Get or Set Parts of an Object}
\usage{
get_method(x)

\S4method{labels}{DiversityIndex}(object, ...)

\S4method{labels}{RarefactionIndex}(object, ...)

\S4method{get_method}{DiversityIndex}(x)
}
\arguments{
\item{object, x}{An \R object from which to get or set element(s).}

\item{...}{Currently not used.}
}
\value{
\itemize{
\item \code{labels()} returns a suitable set of labels from an object for use in
printing or plotting.
}
}
\description{
Getters and setters to extract or replace parts of an object.
}
\seealso{
Other mutators: 
\code{\link{data.frame}}
}
\author{
N. Frerebeau
}
\concept{mutators}
