% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/extract.R
\docType{class}
\name{PermutationOrder-class}
\alias{PermutationOrder-class}
\alias{[[,PermutationOrder-method}
\title{Permutation order}
\usage{
\S4method{[[}{PermutationOrder}(x, i)
}
\arguments{
\item{x}{A \code{PermutationOrder} object from which to extract element(s).}

\item{i}{A \code{\link{character}} string specifying the element to extract.
Character vectors will be matched to the name of the slots.}
}
\description{
An S4 class to represent a permutation order.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[[}: Extracts informations from a slot selected by
subscript \code{i}.
}}

\section{Slots}{

\describe{
\item{\code{rows}}{An \code{\link{integer}} vector giving the rows permutation.}

\item{\code{columns}}{An \code{\link{integer}} vector giving the columns permutation.}

\item{\code{seriation}}{A \code{\link{character}} string indicating the seriation
method used.}
}}

\author{
N. Frerebeau
}
