% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PermutationOrder-class}
\alias{PermutationOrder-class}
\alias{.PermutationOrder}
\title{Permutation Order}
\description{
An S4 class to represent a permutation order.
}
\section{Slots}{

\describe{
\item{\code{rows}}{An \code{\link{integer}} vector giving the rows permutation.}

\item{\code{columns}}{An \code{\link{integer}} vector giving the columns permutation.}

\item{\code{method}}{A \code{\link{character}} string indicating the seriation method used.}
}}

\section{Subset}{

In the code snippets below, \code{x} is a \code{PermutationOrder} object.
\describe{
\item{\code{x[[i]]}}{Extracts information from a slot selected by subscript \code{i}.
\code{i} is a length-one \code{\link{character}} vector. Returns the corresponding slot
values.}
}
}

\seealso{
Other class: 
\code{\link{DiversityIndex}},
\code{\link{RefineCA-class}}
}
\author{
N. Frerebeau
}
\concept{class}
