% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PermutationOrder-class}
\alias{PermutationOrder-class}
\alias{.PermutationOrder}
\title{Permutation Order}
\description{
An S4 class to represent a permutation order.
}
\section{Slots}{

\describe{
\item{\code{rows}}{An \code{\link{integer}} vector giving the rows permutation.}

\item{\code{columns}}{An \code{\link{integer}} vector giving the columns permutation.}

\item{\code{method}}{A \code{\link{character}} string indicating the seriation
method used.}
}}

\note{
Numeric values are coerced to \code{\link{integer}} as by
 \code{\link[base]{as.integer}} (and hence truncated towards zero).
}
\section{Subset}{

 \describe{
  \item{\code{x[[i]]}}{Extracts informations from a slot selected by
  subscript \code{i}. \code{i} should be one of "\code{id}",
  "\code{rows}", "\code{columns}" or "\code{method}". Any unambiguous
  substring can be given.}
 }
}

\seealso{
Other class: 
\code{\link{DateEvent-class}},
\code{\link{DateMCD-class}},
\code{\link{DiversityIndex}},
\code{\link{IncrementTest-class}},
\code{\link{RefineCA-class}}
}
\author{
N. Frerebeau
}
\concept{class}
