% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot_diversity.R
\docType{methods}
\name{plot_diversity}
\alias{plot_diversity}
\alias{autoplot,DiversityIndex-method}
\alias{plot,DiversityIndex,missing-method}
\title{Diversity Plot}
\usage{
\S4method{autoplot}{DiversityIndex}(object, ...)

\S4method{plot}{DiversityIndex,missing}(x, y, ...)
}
\arguments{
\item{object, x}{A \linkS4class{DiversityIndex} object to be plotted.}

\item{...}{Currently not used.}

\item{y}{Currently not used.}
}
\value{
\itemize{
\item \code{autoplot()} returns a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
Diversity Plot
}
\examples{
\donttest{
## Coerce data to a count matrix
data("chevelon", package = "folio")
chevelon <- as_count(chevelon)

## Assemblage diversity size comparison
## Warning: this may take a few seconds!
idx_heterogeneity <- index_heterogeneity(chevelon, method = "shannon")
sim_heterogeneity <- simulate(idx_heterogeneity)
plot(sim_heterogeneity)

idx_richness <- index_richness(chevelon, method = "none")
sim_richness <- simulate(idx_richness)
plot(sim_richness)
}
}
\seealso{
\code{\link[=index_heterogeneity]{index_heterogeneity()}}, \code{\link[=index_evenness]{index_evenness()}}, \code{\link[=index_richness]{index_richness()}}

Other plot: 
\code{\link{plot_bar}},
\code{\link{plot_line}},
\code{\link{plot_matrix}},
\code{\link{plot_spot}()}
}
\author{
N. Frerebeau
}
\concept{plot}
