% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plotBar}
\alias{plotBar}
\alias{plotBar-method}
\alias{plotBar,CountMatrix-method}
\alias{plotBar,FrequencyMatrix-method}
\title{Bar plot}
\usage{
plotBar(object, ...)

\S4method{plotBar}{CountMatrix}(object, EPPM = FALSE, center = TRUE,
  horizontal = FALSE)

\S4method{plotBar}{FrequencyMatrix}(object, EPPM = FALSE,
  center = TRUE, horizontal = FALSE)
}
\arguments{
\item{object}{An object to be plotted.}

\item{...}{Further arguments passed to other methods.}

\item{EPPM}{A \code{\link{logical}} scalar: should TODO?.}

\item{center}{A \code{\link{logical}} scalar: should the bar plot
be centered? The default value, \code{TRUE}, produces a Ford diagram.}

\item{horizontal}{A \code{\link{logical}} scalar: should the bar plot
be horizontal? The default value, \code{FALSE}, means variables in rows and
cases in columns (i.e. Bertin diagram). Only used if \code{center} is
\code{FALSE}.}
}
\description{
XXX
}
\details{
TODO
 coerce to frequency matrix first
}
\examples{
# Coerce dataset to abundance matrix
# Data from Desachy 2004
count <- as(compiegne, "CountMatrix")
freq <- as(boves, "FrequencyMatrix")

# Plot Ford diagram...
## ...without threshold
plotBar(count)
plotBar(freq)

## ...with EPPM as threshold (i.e. Bruno Desachy's sériographe)
plotBar(count, EPPM = TRUE)
plotBar(freq, EPPM = TRUE)

# Plot Bertin diagram...
## ...without threshold
plotBar(count, center = FALSE, horizontal = TRUE)
plotBar(freq, center = FALSE, horizontal = FALSE)

## ...with EPPM as threshold (i.e. Bruno Desachy's sériographe)
plotBar(count, EPPM = TRUE, center = FALSE, horizontal = TRUE)
}
\references{
Desachy, B. (2004). Le sériographe EPPM: un outil informatisé de sériation
 graphique pour tableaux de comptages. \emph{Revue archéologique de
 Picardie}, 3(1), 39-56.
 DOI: \href{https://doi.org/10.3406/pica.2004.2396}{10.3406/pica.2004.2396}.
}
\seealso{
Other plot: \code{\link{plotMatrix}},
  \code{\link{plotRank}}, \code{\link{plotSpot}}
}
\author{
N. Frerebeau
}
\concept{plot}
