% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plotDate}
\alias{plotDate}
\alias{plotDate-method}
\alias{plotDate,CountMatrix,DateEvent-method}
\alias{plotBar,DateEvent-method}
\title{Date plot}
\usage{
plotDate(object, event, ...)

\S4method{plotDate}{CountMatrix,DateEvent}(object, event,
  type = c("event", "accumulation"), select = 1)

\S4method{plotBar}{DateEvent}(object, select = 1, sort = "dsc")
}
\arguments{
\item{object}{An object to be plotted.}

\item{event}{An object of class \linkS4class{DateEvent} to be plotted.}

\item{...}{Further arguments passed to other methods.}

\item{type}{A \code{\link{character}} string or vector of character strings
indicating the TODO.}

\item{select}{A \code{\link{numeric}} or \code{\link{character}} vector
giving the selection of the elements that are drawn.}

\item{sort}{A \code{\link{character}} string indicating if TODO.
If \code{NULL} TODO.}
}
\description{
Date plot
}
\references{
Bellanger, L. & Husi, P. (2012). Statistical Tool for Dating and
 Interpreting Archaeological Contexts Using Pottery. \emph{Journal of
 Archaeological Science}, 39(4), 777-790.
 DOI: \href{https://doi.org/10.1016/j.jas.2011.06.031}{10.1016/j.jas.2011.06.031}.
}
\seealso{
Other plot: \code{\link{plotBar}},
  \code{\link{plotMatrix}}, \code{\link{plotRank}},
  \code{\link{plotSpot}}
}
\author{
N. Frerebeau
}
\concept{plot}
