% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plotDate}
\alias{plotDate}
\alias{plotDate-method}
\alias{plotDate,DateEvent,missing-method}
\alias{plotDate,missing,DateAccumulation-method}
\alias{plotDate,DateEvent,DateAccumulation-method}
\title{Date plot}
\usage{
plotDate(event, accumulation, ...)

\S4method{plotDate}{DateEvent,missing}(event, assemblage = 1)

\S4method{plotDate}{missing,DateAccumulation}(accumulation,
  assemblage = 1)

\S4method{plotDate}{DateEvent,DateAccumulation}(event, accumulation,
  assemblage = 1)
}
\arguments{
\item{event}{An object of class \linkS4class{DateEvent} to be plotted.}

\item{accumulation}{An object of class \linkS4class{DateAccumulation} to be
plotted.}
}
\description{
Date plot
}
\references{
Bellanger, L. & Husi, P. (2012). Statistical Tool for Dating and
 Interpreting Archaeological Contexts Using Pottery. \emph{Journal of
 Archaeological Science}, 39(4), 777-790.
 DOI: \href{https://doi.org/10.1016/j.jas.2011.06.031}{10.1016/j.jas.2011.06.031}.
}
\seealso{
Other plot: \code{\link{plotBar}},
  \code{\link{plotMatrix}}, \code{\link{plotRank}},
  \code{\link{plotSpot}}
}
\author{
N. Frerebeau
}
\concept{plot}
