% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-date.R, R/plot-time.R
\docType{methods}
\name{plotDate}
\alias{plotDate}
\alias{plotDate-method}
\alias{plotTime}
\alias{plotTime-method}
\alias{plotDate,AbundanceMatrix-method}
\alias{plotDate,DateModel-method}
\alias{plotTime,CountMatrix-method}
\title{Date and Time Plot}
\usage{
plotDate(object, ...)

plotTime(object, ...)

\S4method{plotDate}{AbundanceMatrix}(object, select = NULL,
  sort = "dsc")

\S4method{plotDate}{DateModel}(object, type = c("activity", "tempo"),
  event = FALSE, select = 1, n = 500)

\S4method{plotTime}{CountMatrix}(object, highlight = NULL,
  level = 0.95, roll = FALSE, window = 5, facet = TRUE, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{DateModel} to be plotted.}

\item{...}{Further arguments to be passed to internal methods.}

\item{select}{A \code{\link{numeric}} or \code{\link{character}} vector
giving the selection of the assemblage that are drawn.}

\item{sort}{A \code{\link{character}} string indicating whether the dates
should be sorted. It can be one of "\code{asc}" or "\code{dsc}" (default).
Any unambiguous substring can be given. If \code{NULL} no sorting is
performed.}

\item{type}{A \code{\link{character}} string indicating the type of plot.
It must be one of "\code{activity}" (default) or "\code{tempo}".
Any unambiguous substring can be given.}

\item{event}{A \code{\link{logical}} scalar: should the distribution of the
event date be displayed? Only used if type is "\code{activity}".}

\item{n}{A length-one non-negative \code{\link{numeric}} vector giving the
desired length of the vector of quantiles for density computation.}

\item{highlight}{A \code{\link{character}} string indicating the type of
plot. It must be one of "\code{FIT}" or \code{NULL} (default).
Any unambiguous substring can be given.}

\item{level}{A single \code{\link{numeric}} value giving the XXX.}

\item{roll}{A \code{\link{logical}} scalar: should XXX?
Only used if \code{type} is "\code{FIT}" (see details).}

\item{window}{An odd \code{\link{integer}} giving the size of the rolling
window. Only used if \code{roll} is \code{TRUE}.}

\item{facet}{A \code{\link{logical}} scalar: should a matrix of panels
defined by type/taxon be drawn? Only used if XXX.}
}
\value{
TODO
}
\description{
\code{plotDate} produces an activity or tempo plot.
\code{plotTime} produces a time \emph{vs} abundance diagram.
}
\details{
Plots the two probability estimate density curves of
 archaeological assembalge dates (\emph{event} and
 \emph{accumulation} dates; Bellanger and Husi 2012).

 The estimated probability density of an event date is approached by a normal
 distribution. The distribution of the accumulation time of each context is
 approached by a Gaussian mixture.

 The \emph{event date} is plotted as a line, while the \emph{accumulation
 time} is shown as a grey filled area.
}
\section{Detection of selective processes}{

 TODO
}

\references{
Bellanger, L. & Husi, P. (2012). Statistical Tool for Dating and
 Interpreting Archaeological Contexts Using Pottery. \emph{Journal of
 Archaeological Science}, 39(4), 777-790.
 DOI: \href{https://doi.org/10.1016/j.jas.2011.06.031}{10.1016/j.jas.2011.06.031}.

 Dye, T. S. (2016). Long-Term Rhythms in the Development of Hawaiian
 Social Stratification. \emph{Journal of Archaeological Science}, 71, 1-9.
 DOI: \href{https://doi.org/10.1016/j.jas.2016.05.006}{10.1016/j.jas.2016.05.006}.
}
\seealso{
\link{dateEvent}, \link{test}

Other plot: \code{\link{plotBar-method}},
  \code{\link{plotLine}}, \code{\link{plotMatrix}},
  \code{\link{plotSpot}}
}
\author{
N. Frerebeau
}
\concept{plot}
