% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-rank.R
\docType{methods}
\name{plotLine}
\alias{plotLine}
\alias{plotRank}
\alias{plotRank-method}
\alias{plotRank,CountMatrix-method}
\alias{plotRank,FrequencyMatrix-method}
\title{Line plot}
\usage{
plotRank(object, ...)

\S4method{plotRank}{CountMatrix}(object, log = NULL, facet = TRUE)

\S4method{plotRank}{FrequencyMatrix}(object, log = NULL, facet = TRUE)
}
\arguments{
\item{object}{An object to be plotted.}

\item{...}{Further arguments passed to other methods.}

\item{log}{A \code{\link{character}} string which contains "\code{x}" if the
x axis is to be logarithmic, "\code{y}" if the y axis is to be logarithmic
and "\code{xy}" or "\code{yx}" if both axes are to be logarithmic (base 10).}

\item{facet}{A \code{\link{logical}} scalar: should a matrix of panels
defined by case/sample be drawn?}
}
\description{
\code{plotRank} plots a rank \emph{vs} relative abundance diagram.
}
\details{
Note that data are scaled to \eqn{0}-\eqn{1} (relative frequencies).
}
\examples{
## Coerce datasets to abundance matrix
## Data from Desachy 2004
data("compiegne")
count_compiegne <- as(compiegne, "CountMatrix")

## Plot rank vs abundance
plotRank(count_compiegne)
plotRank(count_compiegne, facet = FALSE)
}
\seealso{
Other plot: \code{\link{plotBar}}, \code{\link{plotDate}},
  \code{\link{plotMatrix}}, \code{\link{plotSpot}}
}
\author{
N. Frerebeau
}
\concept{plot}
