% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plotSpot}
\alias{plotSpot}
\alias{plotSpot-method}
\alias{plotSpot,CountMatrix-method}
\alias{plotSpot,FrequencyMatrix-method}
\title{Spot plot}
\usage{
plotSpot(object, ...)

\S4method{plotSpot}{CountMatrix}(object, threshold = NULL)

\S4method{plotSpot}{FrequencyMatrix}(object, threshold = NULL)
}
\arguments{
\item{object}{An object to be plotted.}

\item{...}{Further arguments passed to other methods.}

\item{threshold}{A \code{\link{character}} string specifiying the threshold
to be plotted. This must be one or more of "\code{mean}" or "\code{median}".
Any unambiguous substring can be given.}
}
\description{
XXX
}
\details{
TODO
variables scaled to 0-1 (frequency)
}
\note{
Adapted from Dan Gopstein's original
 \href{https://dgopstein.github.io/articles/spot-matrix/}{spot matrix}.
 Credit should be given to him.
}
\examples{
# Coerce dataset to abundance matrix
# Data from Desachy 2004
count <- as(compiegne, "CountMatrix")
freq <- as(boves, "FrequencyMatrix")

# Plot spot diagram...
## ...without threshod
plotSpot(count)

## ...with column means as threshold
plotSpot(count, threshold = "mean")

# Plot spot diagram with column medians as threshold
plotSpot(freq, threshold = "median")
}
\seealso{
Other plot: \code{\link{plotBar}},
  \code{\link{plotMatrix}}, \code{\link{plotRank}}
}
\author{
N. Frerebeau
}
\concept{plot}
