% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/refine.R
\docType{methods}
\name{refine}
\alias{refine}
\alias{refine-method}
\alias{refine,CountMatrix-method}
\alias{refine,DateModel-method}
\title{Model Refining}
\usage{
refine(object, ...)

\S4method{refine}{CountMatrix}(object, cutoff, n = 1000, axes = c(1,
  2), ...)

\S4method{refine}{DateModel}(object, method = c("jackknife",
  "bootstrap"), n = 1000, ...)
}
\arguments{
\item{object}{An \eqn{m \times p}{m x p} data matrix.}

\item{...}{Further arguments passed to other methods.}

\item{cutoff}{A function that takes a numeric vector as argument and returns
a single numeric value (see details).}

\item{n}{A non-negative \code{\link{integer}} giving the number of partial
bootstrap replications (see details).}

\item{axes}{A \code{\link{numeric}} vector giving the subscripts of the CA
axes to use (see details).}

\item{method}{A \code{\link{character}} string specifiying the resampling
method to be used. This must be one of "\code{jackknife}",
"\code{bootstrap}" (see details). Any unambiguous substring can be given.}
}
\value{
TODO
}
\description{
Model Refining
}
\section{Methods (by class)}{
\itemize{
\item \code{CountMatrix}: Performs a partial bootstrap correspondance analysis
seriation refinement and returns a \linkS4class{BootCA} object.

\item \code{DateModel}: Checks the stability of a date model with resampling
methods.
}}

\note{
These methods can lead to much longer execution times and larger output
 objects. To monitor the execution of these re-sampling procedures, a
 progress bar will automatically be displayed if the
 \code{\link[pbapply]{pbapply}} package is installed on your machine.
}
\section{CA seriation refining}{

 \code{refine} allows to identify samples that are subject to sampling error
 or samples that have underlying structural relationships and might be
 influencing the ordering along the CA space.

 This relies on a partial bootstrap approach to CA-based seriation where each
 sample is replicated \code{n} times. The maximum dimension length of
 the convex hull around the sample point cloud allows to remove samples for
 a given \code{cutoff} value.

 According to Peebles and Schachner (2012), "[this] point removal procedure
 [results in] a reduced dataset where the position of individuals within the
 CA are highly stable and which produces an ordering consistend with the
 assumptions of frequency seriation."

 If the results of \code{\link{refine}} is used as an input argument in
 \code{seriate}, a correspondance analysis is performed on the subset of
 \code{object} which matches the samples to be kept. Then excluded samples
 are projected onto the dimensions of the CA coordinate space using the row
 transition formulae. Finally, row coordinates onto the first dimension
 give the seriation order.
}

\section{Date model checking}{

 Resampling methods can be used to check the stability of the resulting
 \linkS4class{DateModel}.

 If \code{jackknife} is used, one type/fabric is removed at a
 time and all statistics are recalculated. In this way, one can assess
 whether certain type/fabric has a substantial influence on the date
 estimate.
 A six columns \code{\link{data.frame}} is returned, giving the results of
 the resamping procedure (jackknifing fabrics) for each assemblage (in rows)
 with the following columns:
 \describe{
  \item{id}{An identifier to link each row to an assemblage.}
  \item{date}{The jackknife event date estimate.}
  \item{lower}{The lower boundary of the associated prediction interval.}
  \item{upper}{The upper boundary of the associated prediction interval.}
  \item{error}{The standard error of predicted means.}
  \item{bias}{The jackknife estimate of bias.}
 }

 If \code{bootstrap} is used, a large number of new
 bootstrap assemblages is created, with the same sample size, by resampling
 each of the original assemblage with replacement. Then, examination of the
 bootstrap statistics makes it possible to pinpoint assemblages that require
 further investigation.
 A six columns \code{\link{data.frame}} is returned, giving the boostrap
 distribution statistics for each replicated assemblage (in rows)
 with the following columns:
 \describe{
  \item{id}{An identifier to link each row to an assemblage.}
  \item{min}{Minimum value.}
  \item{Q05}{Sample quantile to 0.05 probability.}
  \item{mean}{Mean value (event date).}
  \item{Q95}{Sample quantile to 0.95 probability.}
  \item{max}{Maximum value.}
 }
}

\references{
Peeples, M. A., & Schachner, G. (2012). Refining correspondence
 analysis-based ceramic seriation of regional data sets. \emph{Journal of
 Archaeological Science}, 39(8), 2818-2827.
 DOI: \href{https://doi.org/10.1016/j.jas.2012.04.040}{10.1016/j.jas.2012.04.040}.
}
\seealso{
\link{seriate}, \link{date}
}
