% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/refine_ca.R, R/refine_dates.R,
%   R/refine_diversity.R
\docType{methods}
\name{refine}
\alias{refine}
\alias{bootstrap}
\alias{bootstrap-method}
\alias{jackknife}
\alias{jackknife-method}
\alias{bootstrap,CA-method}
\alias{jackknife,DateModel-method}
\alias{bootstrap,DateModel-method}
\alias{bootstrap,DateMCD-method}
\alias{bootstrap,DiversityIndex-method}
\alias{jackknife,DiversityIndex-method}
\alias{simulate,DiversityIndex-method}
\title{Bootstrap and Jackknife Resampling}
\usage{
bootstrap(object, ...)

jackknife(object, ...)

\S4method{bootstrap}{CA}(
  object,
  cutoff,
  n = 1000,
  axes = c(1, 2),
  progress = getOption("tabula.progress"),
  ...
)

\S4method{jackknife}{DateModel}(object, level = 0.95, progress = getOption("tabula.progress"), ...)

\S4method{bootstrap}{DateModel}(
  object,
  level = 0.95,
  probs = c(0.05, 0.95),
  n = 1000,
  progress = getOption("tabula.progress"),
  ...
)

\S4method{bootstrap}{DateMCD}(object, probs = c(0.05, 0.95), n = 1000, ...)

\S4method{bootstrap}{DiversityIndex}(object, probs = c(0.05, 0.95), n = 1000)

\S4method{jackknife}{DiversityIndex}(object)

\S4method{simulate}{DiversityIndex}(
  object,
  quantiles = TRUE,
  level = 0.8,
  step = 1,
  n = 1000,
  progress = getOption("tabula.progress"),
  ...
)
}
\arguments{
\item{object}{A \linkS4class{CA}, \linkS4class{DateModel} or
\linkS4class{DiversityIndex} object.}

\item{...}{Currently not used.}

\item{cutoff}{A function that takes a numeric vector as argument and returns
a single numeric value (see below).}

\item{n}{A non-negative \code{\link{integer}} giving the number of bootstrap
replications (see below).}

\item{axes}{A \code{\link{numeric}} vector giving the subscripts of the CA
axes to be used (see below).}

\item{progress}{A \code{\link{logical}} scalar: should a progress bar be
displayed?}

\item{level}{A length-one \code{\link{numeric}} vector giving the
confidence level.}

\item{probs}{A \code{\link{numeric}} vector of probabilities with values in
\eqn{[0,1]} (see \code{\link[stats:quantile]{quantile}}).}

\item{quantiles}{A \code{\link{logical}} scalar: should sample quantiles
be used as confidence interval? If \code{TRUE} (the default),
sample quantiles are used as described in Kintigh (1989), else quantiles of
the normal distribution are used.}

\item{step}{A non-negative \code{\link{integer}} giving the increment of the
sample size. Only used if \code{simulate} is \code{TRUE}.}
}
\value{
\code{bootstrap} and \code{jackknife} return a \code{\link{data.frame}}.

 \code{bootstrap} returns a \linkS4class{BootCA} object.
}
\description{
Checks model with resampling methods.
}
\section{Functions}{
\itemize{
\item \code{bootstrap,CA-method}: Performs a partial bootstrap correspondence analysis.

\item \code{jackknife,DateModel-method}: Jackknifes fabrics.

\item \code{bootstrap,DateModel-method}: Bootstraps assemblages.

\item \code{bootstrap,DateMCD-method}: Bootstrap Mean Ceramic Date confidence intervals.

\item \code{bootstrap,DiversityIndex-method}: TODO.

\item \code{jackknife,DiversityIndex-method}: TODO.

\item \code{simulate,DiversityIndex-method}: TODO.
}}

\note{
Refining method can lead to much longer execution times and larger output
 objects. To monitor the execution of these re-sampling procedures, a
 progress bar will be displayed.
}
\section{Correspondence Analysis}{

 \code{bootstrap} allows to identify samples that are subject to
 sampling error or samples that have underlying structural relationships
 and might be influencing the ordering along the CA space.

 This relies on a partial bootstrap approach to CA-based seriation where each
 sample is replicated \code{n} times. The maximum dimension length of
 the convex hull around the sample point cloud allows to remove samples for
 a given \code{cutoff} value.

 According to Peebles and Schachner (2012), "[this] point removal procedure
 [results in] a reduced dataset where the position of individuals within the
 CA are highly stable and which produces an ordering consistent with the
 assumptions of frequency seriation."
}

\section{Date Model}{

 If \code{jackknife} is used, one type/fabric is removed at a
 time and all statistics are recalculated. In this way, one can assess
 whether certain type/fabric has a substantial influence on the date
 estimate.
 A six columns \code{\link{data.frame}} is returned, giving the results of
 the resampling procedure (jackknifing fabrics) for each assemblage (in rows)
 with the following columns:
 \describe{
  \item{id}{An identifier to link each row to an assemblage.}
  \item{date}{The jackknife event date estimate.}
  \item{lower}{The lower boundary of the associated prediction interval.}
  \item{upper}{The upper boundary of the associated prediction interval.}
  \item{error}{The standard error of predicted means.}
  \item{bias}{The jackknife estimate of bias.}
 }

 If \code{bootstrap} is used, a large number of new
 bootstrap assemblages is created, with the same sample size, by resampling
 each of the original assemblage with replacement. Then, examination of the
 bootstrap statistics makes it possible to pinpoint assemblages that require
 further investigation.

 A five columns \code{\link{data.frame}} is returned, giving the bootstrap
 distribution statistics for each replicated assemblage (in rows)
 with the following columns:
 \describe{
  \item{min}{Minimum value.}
  \item{mean}{Mean value (event date).}
  \item{max}{Maximum value.}
  \item{Q5}{Sample quantile to 0.05 probability.}
  \item{Q95}{Sample quantile to 0.95 probability.}
 }
}

\examples{
## Data from Magurran 1988, p. 145-149
birds <- CountMatrix(
  data = c(35, 26, 25, 21, 16, 11, 6, 5, 3, 3,
           3, 3, 3, 2, 2, 2, 1, 1, 1, 1, 0, 0,
           30, 30, 3, 65, 20, 11, 0, 4, 2, 14,
           0, 3, 9, 0, 0, 5, 0, 0, 0, 0, 1, 1),
  nrow = 2, byrow = TRUE, dimnames = list(c("oakwood", "spruce"), NULL))

## Shannon diversity
h <- index_heterogeneity(birds, "shannon")
bootstrap(h)
jackknife(h)

## Shannon evenness
e <- index_evenness(birds, "shannon")
bootstrap(e)
jackknife(e)
}
\references{
Bellanger, L., Tomassone, R. & Husi, P. (2008). A Statistical Approach for
 Dating Archaeological Contexts. \emph{Journal of Data Science}, 6, 135-154.

 Peeples, M. A., & Schachner, G. (2012). Refining correspondence
 analysis-based ceramic seriation of regional data sets. \emph{Journal of
 Archaeological Science}, 39(8), 2818-2827. \doi{10.1016/j.jas.2012.04.040}.
}
\seealso{
\link{date_event}, \link{index_heterogeneity},
 \link{index_evenness}, \link{index_richness}, \link{seriate_correspondence}

Other statistics: 
\code{\link{independance}},
\code{\link{test_diversity}()},
\code{\link{test_fit}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
