% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{refine}
\alias{refine}
\alias{refine-method}
\title{Refine models}
\usage{
refine(object, ...)
}
\arguments{
\item{object}{An object to refine.}

\item{...}{Further arguments passed to other methods.}

\item{axes}{A \code{\link{numeric}} vector giving the subscripts of the CA
axes to use (see details).}

\item{cutoff}{A function that takes a numeric vector as argument and returns
a single numeric value (see details).}

\item{bootstrap}{A \code{\link{logical}} scalar: should the model be checked
by bootstrap resampling?}

\item{jackknife}{A \code{\link{logical}} scalar: should the model be checked
by jackknife estimation (removing each fabric/type one at a time)?}

\item{n}{A non-negative \code{\link{integer}} giving the number of partial
bootstrap replications (see details).}
}
\description{
See references bellow for more details.
}
\references{
Bellanger, L., Husi, P. & Tomassone, R. (2006). Une approche statistique
 pour la datation de contextes archéologiques. \emph{Revue de Statistique
 Appliquée}, 54(2), 65-81.

 Bellanger, L., Husi, P. & Tomassone, R. (2006). Statistical Aspects of
 Pottery Quantification for the Dating of Some Archaeological Contexts.
 \emph{Archaeometry}, 48(1), 169-183.
 DOI: \href{https://doi.org/10.1111/j.1475-4754.2006.00249.x}{10.1111/j.1475-4754.2006.00249.x}.

 Peeples, M. A., & Schachner, G. (2012). Refining correspondence
 analysis-based ceramic seriation of regional data sets. \emph{Journal of
 Archaeological Science}, 39(8), 2818-2827.
 DOI: \href{https://doi.org/10.1016/j.jas.2012.04.040}{10.1016/j.jas.2012.04.040}.
}
\seealso{
\link{dateEvent} \link{seriate}
}
\author{
N. Frerebeau
}
