% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/index_diversity.R
\docType{methods}
\name{resample}
\alias{resample}
\alias{bootstrap,DiversityIndex-method}
\alias{jackknife,DiversityIndex-method}
\title{Resample Diversity}
\usage{
\S4method{bootstrap}{DiversityIndex}(object, n = 1000, f = NULL)

\S4method{jackknife}{DiversityIndex}(object)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector of count data (absolute frequencies).}

\item{n}{A non-negative \code{\link{integer}} giving the number of bootstrap
replications.}

\item{f}{A \code{\link{function}} that takes a single numeric vector (the result of
\code{do}) as argument.}
}
\value{
Returns a \code{\link{data.frame}}.
}
\description{
Performs bootstrap/jackknife resampling.
}
\examples{
data("chevelon", package = "folio")

## Shannon diversity index
(h <- heterogeneity(chevelon, method = "shannon"))
(e <- evenness(chevelon, method = "shannon"))

## Bootstrap resampling (summary statistics)
bootstrap(h, f = summary)

## Jackknife resampling
jackknife(h)
}
\seealso{
Other diversity methods: 
\code{\link{heterogeneity}()},
\code{\link{rarefaction}()},
\code{\link{richness}()},
\code{\link{simulate}()},
\code{\link{turnover}()}
}
\author{
N. Frerebeau
}
\concept{diversity methods}
