% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SpaceTime-class}
\alias{SpaceTime-class}
\title{Space and Time}
\description{
An S4 class to reprensent space-time informations.
}
\section{Slots}{

\describe{
\item{\code{dates}}{A list of two \code{\link{numeric}} vectors giving
the mean date and error, respectively.}

\item{\code{coordinates}}{A list of three \code{\link{numeric}} vectors
giving the geographic coordinates (longitude, latitude and elevation,
respectively).}

\item{\code{epsg}}{An \code{\link{integer}} giving the EPSG code of the spatial
reference system used. Numeric values are coerced to \code{\link{integer}}
as by \code{\link{as.integer}} (and hence truncated towards zero).}
}}

\author{
N. Frerebeau
}
\keyword{internal}
