% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stats_jackknife}
\alias{stats_jackknife}
\title{Jackknife Estimation}
\usage{
stats_jackknife(x, do, ...)
}
\arguments{
\item{x}{A vector.}

\item{do}{A \code{\link{function}} that takes \code{x} as an argument
and returns a single numeric value.}

\item{...}{Extra arguments passed to \code{do}.}
}
\value{
A \code{\link{numeric}} vector with the following elements:
\describe{
\item{\code{values}}{The \eqn{n} leave-one-out values.}
\item{\code{mean}}{The jackknife estimate of mean.}
\item{\code{bias}}{The jackknife estimate of bias.}
\item{\code{error}}{he jackknife estimate of standard error.}
}
}
\description{
Jackknife Estimation
}
\keyword{internal}
