% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/extract.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{[,BootCA-method}
\alias{[,DateModel-method}
\alias{[[,DateModel-method}
\alias{[[,BootCA-method}
\alias{[[,PermutationOrder-method}
\alias{[[,SpaceTime-method}
\alias{[[,CountMatrix-method}
\alias{[[,FrequencyMatrix-method}
\alias{[[,IncidenceMatrix-method}
\alias{[[,OccurrenceMatrix-method}
\alias{[[,SimilarityMatrix-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[}{BootCA}(x, i, j, drop = TRUE)

\S4method{[}{DateModel}(x, i, j, drop = TRUE)

\S4method{[[}{DateModel}(x, i)

\S4method{[[}{BootCA}(x, i)

\S4method{[[}{PermutationOrder}(x, i)

\S4method{[[}{SpaceTime}(x, i)

\S4method{[[}{CountMatrix}(x, i)

\S4method{[[}{FrequencyMatrix}(x, i)

\S4method{[[}{IncidenceMatrix}(x, i)

\S4method{[[}{OccurrenceMatrix}(x, i)

\S4method{[[}{SimilarityMatrix}(x, i)
}
\arguments{
\item{x}{An object from which to extract element(s) or in which to replace
element(s).}

\item{i, j}{Indices specifying elements to extract or replace. Indices are
\code{\link{numeric}}, \code{\link{integer}} or \code{\link{character}}
vectors or empty (missing) or \code{NULL}. Numeric values are coerced to
\code{\link{integer}} as by \code{\link{as.integer}} (and hence truncated
towards zero). Character vectors will be matched to the name of the
elements. An empty index (a comma separated blank) indicates that all
entries in that dimension are selected.}

\item{drop}{A \code{\link{logical}} scalar: should the result be coerced to
the lowest possible dimension? This only works for extracting elements,
not for the replacement.}
}
\value{
TODO
}
\description{
Operators acting on objects to extract or replace parts.
}
\author{
N. Frerebeau
}
