% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/subset.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{[[,DiversityIndex,ANY,missing-method}
\alias{[[,IncrementTest,ANY,missing-method}
\alias{[[,DateMCD,ANY,missing-method}
\alias{[[,DateEvent,ANY,missing-method}
\alias{[[,RefineCA,ANY,missing-method}
\alias{[[,PermutationOrder,ANY,missing-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[[}{DiversityIndex,ANY,missing}(x, i)

\S4method{[[}{IncrementTest,ANY,missing}(x, i)

\S4method{[[}{DateMCD,ANY,missing}(x, i)

\S4method{[[}{DateEvent,ANY,missing}(x, i)

\S4method{[[}{RefineCA,ANY,missing}(x, i)

\S4method{[[}{PermutationOrder,ANY,missing}(x, i)
}
\arguments{
\item{x}{An object from which to extract element(s) or in which to replace
element(s).}

\item{i}{A \code{\link{character}} string specifying elements to extract.
Any unambiguous substring can be given (see details).}
}
\value{
A subsetted object.
}
\description{
Operators acting on objects to extract or replace parts.
}
\seealso{
Other mutator: 
\code{\link{mutator}}
}
\author{
N. Frerebeau
}
\concept{mutator}
