% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{subset}
\alias{subset}
\title{Extract or Replace Parts of an Object}
\arguments{
\item{x}{An object from which to extract element(s) or in which to replace
element(s).}

\item{i}{A \code{\link{character}} string specifying elements to extract.
Any unambiguous substring can be given (see details).}
}
\value{
A subsetted object.
}
\description{
Operators acting on objects to extract or replace parts.
}
\seealso{
Other mutator: 
\code{\link{mutator}}
}
\author{
N. Frerebeau
}
\concept{mutator}
