% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabula-deprecated.R
\name{tabula-deprecated}
\alias{tabula-deprecated}
\title{Deprecated Functions in tabula}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\keyword{internal}
