% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabula.R
\docType{package}
\name{tabula-package}
\alias{tabula-package}
\alias{tabula}
\title{Analysis, Seriation and Visualization of Archaeological Count Data}
\description{
An easy way to examine archaeological count data (artifacts, faunal remains,
etc.). This package includes several measures of diversity, e.g. richness,
rarefaction, diversity, turnover, similarity, etc. It also provides matrix
seriation methods for chronological modeling and dating. The package make it
easy to visualize count data and statistical thresholds: rank/abundance
plots, Ford and Bertin diagrams, etc.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab tabula \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 0.9.0 \cr
 \strong{Date:} \tab 2018-11-16 \cr
 \strong{License:} \tab GPL-3
}
}
\author{
\strong{Full list of authors and contributors} (alphabetic order)
 \tabular{ll}{
  Nicolas Frerebeau \tab \emph{IRAMAT-CRP2A, Université Bordeaux Montaigne, France} \cr
  Brice Lebrun \tab \emph{CEREGE, Aix-Marseille Université, France}
 }

\strong{Bug reporting}

- \url{https://github.com/nfrerebeau/tabula/issues}

\strong{Package maintainer}

Nicolas Frerebea, IRAMAT-CRP2A, Universite Bordeaux Montaigne, Pessac, France,\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}
}
